### 2024-07-17 #6999 @kthyng
- [ ] Clear research application/in scope

Also @wangyji please comment further on the paper you say is simultaneously under review about this software. Can you post a link to it or is it not available publicly? If that is the case, please describe what is in the paper to compare it with this submission.
