### 2024-07-17 2c9dfbc
Update train_syndata.py
