### 2024-07-17 ef5f3bb
Update train_syndata.py
