### 2024-07-17 8880670
Update train_syndata.py
