### 2024-07-18 #6999 @kthyng
Hi @wangyji I have some suggestions and questions:

* is there a reason you haven't made a release for your package on PyPI or conda-forge?
* What materials would you call docs?
* How would a user verify the software is running properly once they've set it up?
* fix DOIs
* Shorten your paper by moving "Algorithm Features", "Examples", and "Mathematical Formulation" (or some of those sections if you really want to keep one in the paer) to an actual docs like readthedocs, which you could also add your examples and tutorials to which can then have static versions hosted.
