+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = [
 ]
 description = "A differentiable neural-network solver for data assimilation of ice shelves written in JAX"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
