+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "DIFFICE_jax"
-version = "0.0.1"
+version = "0.0.2"
 authors = [
   { name="Yongji Wang", email="yw8211@nyu.edu" },
 ]
