### 2024-07-26 #6999 @wangyji
> Hi @wangyji, please get back to me about these questions so we can move forward with your submission processing. I'd like to hear back from you next week otherwise I will withdraw your submission the following week.

Hi, I apologize for the late reply. I am still in the process of addressing all the questions and suggestions you have provided. Below are my answers to your questions:

- is there a reason you haven't made a release for your package on PyPI or conda-forge?

Actually, this is my first time writing a formal Python package, and I am not familiar with how to release a package on PyPI. I am currently learning how to do that, but it is proving to be more challenging than I expected. It might take me some time to get it working properly. I have also realized that I need to restructure my repository to make it compatible.

-What materials would you call docs?

The main folder and each subfolder in my GitHub package have README.md files, which I consider as docs
.
-How would a user verify the software is running properly once they've set it up?

I have included a tutorial example in the repository. My thought is for users to verify the proper installation of the software by running that example.

- fix DOIs

I will address this shortly

- Shorten your paper by moving "Algorithm Features", "Examples", and "Mathematical Formulation" (or some of those sections if you really want to keep one in the paer) to an actual docs like readthedocs, which you could also add your examples and tutorials to which can then have static versions hosted.

That's a good suggestion. I will also address this shortly.
