### 2024-08-04 33fefc6
Update Isotropic.md
