+++ paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-The prediction of ice dynamics relies on knowledge of its viscosity structure, which can not be directly measured at the Antarctic scale. Misrepresenting viscsoity in ice-dynamics simulation can lead to imprecise forecasts of ice sheet's mass loss into the oceans and its consequential impact on global sea-level rise. With the continent-wide remote-sensing data available over the past decades, the viscosity of the ice shelves can be inferred by solving an inverse problem. We present `DIFFICE_jax`: a PINNs-based DIFFerentiable solver for data assimilation and inverse modeling of ICE shelves in JAX. This python package converts descretized remote-sensing data into meshless and differentible functions, and infer the viscosity profile by directly solving Shallow-Shelf Approximation (SSA) equations for ice shelves using back-propagation and automatic differentiation (AD). The inversion algorithm is implemented in JAX [@jax2018github]. The `DIFFICE_jax` package involves several advanced features beyond vanilla PINNs algorithms, including collocation points resampling, non-dimensionalization of the data adnd equations, extended PINN, viscosity exponential scaling function, which are essential for accurate inversion. The package is designed to be user-friendly and accessible for beginners. The Github respository also provides tutorial examples with Colab notebooks for users at different levels to reproduce the result and modify the code for their problems of interests.
+The flowing of Antarctica ice shelves is controlled by their viscosity structure, which can not be directly measured at the continental scale. Misrepresenting viscsoity in ice-dynamics simulation can lead to imprecise forecasts of ice sheet's mass loss into the oceans and its consequential impact on global sea-level rise. With the continent-wide remote-sensing data available over the past decades, the viscosity of the ice shelves can be inferred by solving an inverse problem. We present `DIFFICE_jax`: a PINNs-based DIFFerentiable solver for data assimilation and inverse modeling of ICE shelves in JAX. This python package converts descretized remote-sensing data into meshless and differentible functions, and infer the viscosity profile by directly solving Shallow-Shelf Approximation (SSA) equations for ice shelves using back-propagation and automatic differentiation (AD). The inversion algorithm is implemented in JAX [@jax2018github]. The `DIFFICE_jax` package involves several advanced features beyond vanilla PINNs algorithms, including collocation points resampling, non-dimensionalization of data and equations, extended-PINNs techniques, function design of viscosity, which are essential for accurate inversion. The package is designed to be user-friendly and accessible for beginners. The Github respository also provides tutorial examples with Colab notebooks for users at different levels to reproduce the result and modify the code for their problems of interests.
 
 
 # Statement of need
