+++ paper.md
@@ -32,16 +32,19 @@ The flowing of Antarctica ice shelves is controlled by their viscosity structure
 
 # Statement of need
 
-One of the unsolved questions in the cryosphere, or the broader field of geophysics, is how to uncover the hidden physical properties of various geophysical flows on Earth, such as glacial ice [@Millstein2022ice; @ranganathan2024modified]. Ice shelves are the floating extensions of grounded ice sheet that play a critical role in slowing ice discharging into the ocean, mitigating the global sea level rise. A crtical physical properties required to predict ice dynamics is the effective ice viscosity. However, continent-wide in-situ measurement of the viscosity is challenging if not impossible. With high-resolution remote-sensing data available for both ice velocity $(u, v)$ and thickness $h$ across Antarctica, effective viscosity $\mu$ can be inferred via solving an inverse problem constrained by the governing equation of ice-shelf dynamics, i.e. the Shallow-Shelf Approximation (SSA) equations [@MacAyeal1989].
+One of the unsolved questions in the cryosphere, or the broader field of geophysics, is how to uncover the hidden physical properties of various geophysical flows on Earth, such as ice-shelf flows [@Millstein2022ice; @ranganathan2024modified]. Ice shelves are the floating extensions of grounded ice sheet that play a critical role in slowing ice discharging into the ocean, mitigating the global sea level rise. A crtical physical properties required to predict ice dynamics is the effective ice viscosity. However, continent-wide in-situ measurement of the viscosity is challenging if not impossible. Instead, with high-resolution remote-sensing data available for both ice velocity $(u, v)$ and thickness $h$ across Antarctica, effective viscosity $\mu$ can be inferred via solving an inverse problem constrained by the governing equation of ice-shelf dynamics, i.e. the Shallow-Shelf Approximation (SSA) equations [@MacAyeal1989].
+
+Conventional method for solving inverse problems in glaciology includes control method, or called adjoint method [@MacAyeal1993; @vieli2003application; @morlighem2010spatial; @goldberg2011data; @morlighem2013inversion; @perego2014optimal; @furst2015assimilation; @barnes2021transferability]. It is a PDE-constraint optimization method that requires deriving extra adjoint equations  [@MacAyeal1993; @morlighem2013inversion] for the computation. In addition, inverse problem often involves data with certain level of noise, which requires regularization techniques to avoid error propagation from the data noise to the inversion. The above two challenges create certain obstacles for broad communities to solve inverse problem using adjoint method. In comparison, the newly-developed physics-informed neural networks (PINNs) is easy to employ. It has been demonstrated that PINNs can inherently de-noise the data while solving the inverse problem [@wang2022discovering; @riel2023variational; @iwasaki2023one] without the need of regularization terms in the cost function. PINNs also do not require the derivation of ajoint equations, which is convenient for solving complex problems that involves multiple coupled PDEs. 
+
 
[…]
