+++ paper.md
@@ -27,19 +27,18 @@ bibliography: paper.bib
 
 # Summary
 
-The flowing of Antarctica ice shelves is controlled by their viscosity structure, which can not be directly measured at the continental scale. Misrepresenting viscsoity in ice-dynamics simulation can lead to imprecise forecasts of ice sheet's mass loss into the oceans and its consequential impact on global sea-level rise. With the continent-wide remote-sensing data available over the past decades, the viscosity of the ice shelves can be inferred by solving an inverse problem. We present `DIFFICE_jax`: a PINNs-based DIFFerentiable solver for data assimilation and inverse modeling of ICE shelves in JAX. This python package converts descretized remote-sensing data into meshless and differentible functions, and infer the viscosity profile by directly solving Shallow-Shelf Approximation (SSA) equations for ice shelves using back-propagation and automatic differentiation (AD). The inversion algorithm is implemented in JAX [@jax2018github]. The `DIFFICE_jax` package involves several advanced features beyond vanilla PINNs algorithms, including collocation points resampling, non-dimensionalization of data and equations, extended-PINNs techniques, function design of viscosity, which are essential for accurate inversion. The package is designed to be user-friendly and accessible for beginners. The Github respository also provides tutorial examples with Colab notebooks for users at different levels to reproduce the result and modify the code for their problems of interests.
+The flow of Antarctica ice shelves is controlled by their viscosity structure, which can not be directly measured at the continental scale. Misrepresenting viscsoity in ice-dynamics simulation can lead to imprecise forecasts of ice sheet's mass loss into the oceans and its consequential impact on global sea-level rise. With the continent-wide remote-sensing data available over the past decades, the viscosity of the ice shelves can be inferred by solving an inverse problem. We present `DIFFICE_jax`: a PINNs-based DIFFerentiable solver for data assimilation and inverse modeling of ICE shelves in JAX. This python package converts descretized remote-sensing data into meshless and differentible functions, and infer the viscosity profile by solving Shallow-Shelf Approximation (SSA) equations for ice shelves directly. The inversion algorithm is implemented in JAX [@jax2018github]. The `DIFFICE_jax` package involves several advanced features beyond vanilla PINNs algorithms, including collocation points resampling, non-dimensionalization of data and equations, extended-PINNs techniques, function design of viscosity, which are essential for accurate inversion. The package is designed to be user-friendly and accessible for beginners. The Github respository also provides tutorial examples with Colab notebooks for users at different levels to reproduce the result and modify the code for their problems of interests.
 
 
 # Statement of need
 
 One of the unsolved questions in the cryosphere, or the broader field of geophysics, is how to uncover the hidden physical properties of various geophysical flows on Earth, such as ice-shelf flows [@Millstein2022ice; @ranganathan2024modified]. Ice shelves are the floating extensions of grounded ice sheet that play a critical role in slowing ice discharging into the ocean, mitigating the global sea level rise. A crtical physical properties required to predict ice dynamics is the effective ice viscosity. However, continent-wide in-situ measurement of the viscosity is challenging if not impossible. Instead, with high-resolution remote-sensing data available for both ice velocity $(u, v)$ and thickness $h$ across Antarctica, effective viscosity $\mu$ can be inferred via solving an inverse problem constrained by the governing equation of ice-shelf dynamics, i.e. the Shallow-Shelf Approximation (SSA) equations [@MacAyeal1989].
 
[…]
