### 2024-08-04 c2096a3
Update paper.md
