+++ paper.md
@@ -36,7 +36,7 @@ One of the unsolved questions in the cryosphere, or the broader field of geophys
 
 Conventional method for solving inverse problems in glaciology includes the control method, or called the adjoint method [@MacAyeal1993; @vieli2003application; @morlighem2010spatial; @goldberg2011data; @morlighem2013inversion; @perego2014optimal; @furst2015assimilation; @barnes2021transferability], which is a PDE-constraint optimization method that requires deriving extra adjoint equations [@MacAyeal1993; @morlighem2013inversion]. Data used in inverse problems often contains noise. Thus, the adjoint method requires additional regularization techniques to prevent error propagation. In comparison, it has been demonstrated that the physical-informed neural networks (PINNs) can inherently de-noise the data while solving the inverse problem [@wang2022discovering; @riel2023variational; @iwasaki2023one] without the need of regularization terms in the cost function. Moreover, based on back-propagation and automatic differentiation, PINNs also do not require the derivation of ajoint equations, which is convenient for solving complex problems that involves multiple coupled PDEs. 
 
-The `DIFFICE_jax` package is designed to make the inversion of ice-shelf viscosity via PINNs more accessible for beginners. Basic PINNs code are not adequate for accurately inferring effective viscosity $\mu$ from the governing equations and real-world data. Additional settings for neural networks (NN), optimization methods, and pre-processing of both the SSA equations and observational data are all essential for the success of viscosity inversion via PINNs. The `DIFFICE_jax` package incorperates the optimal settings of PINNs in all these aspects. These settings are either universally applicable for training different ice shelves or can be determined automatically based on the data of given ice shelf. The package provides comprehensive details of the key algorithms involved, including comments and examples, enabling users to bypass substantial hyper-parameter tuning and conduct cutting-edge research in the field.
+The `DIFFICE_jax` package is designed to make the inversion of ice-shelf viscosity via PINNs more accessible for beginners. Off-the-shelf PINNs code are not adequate for accurately inferring effective viscosity $\mu$ from the governing equations and real-world data. Additional settings for neural networks (NN), optimization methods, and pre-processing of both the SSA equations and observational data are all essential for the success of viscosity inversion via PINNs. The `DIFFICE_jax` package incorperates the optimal settings of PINNs in all these aspects. These settings are either universally applicable for training different ice shelves or can be determined automatically based on the data of given ice shelf. The package provides comprehensive details of the key algorithms involved, including comments and examples, enabling users to bypass substantial hyper-parameter tuning and conduct cutting-edge research in the field.
 
 ANother advantage of PINNs is its effectiveness with irregularly sampled data. The `DIFFICE_jax` package supports the direct use of thickness data from radar profiles available only at flight lines. In addition, the outputs of PINNs (velocity, thickness, and viscosity fields) are continuous functions parameterized by a fixed number of weights and biases, requiring less memory than the output of classical methods with discretized grid points when higher resolutions are demanded [@wang2024multi]. Finally, the `DIFFICE_jax` package, along with other deep-learning-based solver for ice-dynamics [@brinkerhoff2021constraining; @riel2021data; @riel2023variational; @wang2022discovering; @iwasaki2023one; @jouvet2022deep; @jouvet2023inversion; @jouvet2023ice; @bolibar2023universal; @he2023hybrid; @cheng2024unified], leverages GPUs for computational speedup. 
 
