+++ paper.bib
@@ -6,7 +6,8 @@
   number={197},
   pages={527--533},
   year={2010},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  DOI={10.3189/002214310792447699}
 }
 
 @article{barnes2021transferability,
@@ -153,7 +154,7 @@
 	volume={3},
 	pages={1--7},
 	year={2022},
-	publisher={Nature Publishing Group}
+	DOI = {10.1038/s43247-022-00385-x}
 }
 
 
