+++ src/figure/backup.bib
@@ -1,0 +1,1378 @@
+@article{arthern2010initialization,
+  title={Initialization of ice-sheet forecasts viewed as an inverse Robin problem},
+  author={Arthern, Robert J and Gudmundsson, G Hilmar},
+  journal={Journal of Glaciology},
+  volume={56},
+  number={197},
+  pages={527--533},
+  year={2010},
+  publisher={Cambridge University Press},
+  DOI={10.3189/002214310792447699}
+}
+
+@article{barnes2021transferability,
+  title={The transferability of adjoint inversion products between different ice flow models},
+  author={Barnes, Jowan M and Dias dos Santos, Thiago and Goldberg, Daniel and Gudmundsson, G Hilmar and Morlighem, Mathieu and De Rydt, Jan},
+  journal={The Cryosphere},
+  volume={15},
+  number={4},
+  pages={1975--2000},
+  year={2021},
+  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+}
+
+@article{perego2014optimal,
+  title={Optimal initial conditions for coupling ice sheet models to Earth system models},
+  author={Perego, Mauro and Price, Stephen and Stadler, Georg},
+  journal={Journal of Geophysical Research: Earth Surface},
+  volume={119},
+  number={9},
+  pages={1894--1917},
+  year={2014},
+  publisher={Wiley Online Library}
+}
+
+@article{cheng2024unified,
+  title={A unified framework for forward and inverse modeling of ice sheet flow using physics-informed neural networks},
+  author={Cheng, Gong and Morlighem, Mathieu and Francis, Sade},
+  journal={Authorea Preprints},
+  year={2024},
+  publisher={Authorea}
+}
+
+@article{brinkerhoff2021constraining,
+  title={Constraining subglacial processes from surface velocity observations using surrogate-based Bayesian inference},
+  author={Brinkerhoff, Douglas and Aschwanden, Andy and Fahnestock, Mark},
+  journal={Journal of Glaciology},
+  volume={67},
+  number={263},
+  pages={385--403},
+  year={2021},
+  publisher={Cambridge University Press}
+}
+
+@article{he2023hybrid,
+  title={A hybrid deep neural operator/finite element method for ice-sheet modeling},
+  author={He, QiZhi and Perego, Mauro and Howard, Amanda A and Karniadakis, George Em and Stinis, Panos},
+  journal={Journal of Computational Physics},
+  volume={492},
+  pages={112428},
+  year={2023},
+  publisher={Elsevier}
+}
+
+@article{bolibar2023universal,
+  title={Universal differential equations for glacier ice flow modelling},
+  author={Bolibar, Jordi and Sapienza, Facundo and Maussion, Fabien and Lguensat, Redouane and Wouters, Bert and P{\'e}rez, Fernando},
+  journal={Geoscientific Model Development Discussions},
+  volume={2023},
+  pages={1--26},
+  year={2023},
+  publisher={G{\"o}ttingen, Germany}
+}
+
+@article{jouvet2022deep,
+  title={Deep learning speeds up ice flow modelling by several orders of magnitude},
+  author={Jouvet, Guillaume and Cordonnier, Guillaume and Kim, Byungsoo and L{\"u}thi, Martin and Vieli, Andreas and Aschwanden, Andy},
+  journal={Journal of Glaciology},
+  volume={68},
+  number={270},
+  pages={651--664},
+  year={2022},
+  publisher={Cambridge University Press}
+}
+
+@article{jouvet2023inversion,
[…]
