+++ paper.bib
@@ -1,15 +1,3 @@
-@article{arthern2010initialization,
-  title={Initialization of ice-sheet forecasts viewed as an inverse Robin problem},
-  author={Arthern, Robert J and Gudmundsson, G Hilmar},
-  journal={Journal of Glaciology},
-  volume={56},
-  number={197},
-  pages={527--533},
-  year={2010},
-  publisher={Cambridge University Press},
-  DOI={10.3189/002214310792447699}
-}
-
 @article{barnes2021transferability,
   title={The transferability of adjoint inversion products between different ice flow models},
   author={Barnes, Jowan M and Dias dos Santos, Thiago and Goldberg, Daniel and Gudmundsson, G Hilmar and Morlighem, Mathieu and De Rydt, Jan},
@@ -18,7 +6,7 @@
   number={4},
   pages={1975--2000},
   year={2021},
-  publisher={Copernicus Publications G{\"o}ttingen, Germany}
+  DOI = {10.5194/tc-15-1975-2021}
 }
 
 @article{perego2014optimal,
@@ -29,7 +17,7 @@
   number={9},
   pages={1894--1917},
   year={2014},
-  publisher={Wiley Online Library}
+  DOI = {https://doi.org/10.1002/2014JF003181}
 }
 
 @article{cheng2024unified,
