+++ paper.bib
@@ -17,7 +17,7 @@
   number={9},
   pages={1894--1917},
   year={2014},
-  DOI = {https://doi.org/10.1002/2014JF003181}
+  DOI = {10.1002/2014JF003181}
 }
 
 @article{cheng2024unified,
@@ -25,7 +25,7 @@
   author={Cheng, Gong and Morlighem, Mathieu and Francis, Sade},
   journal={Authorea Preprints},
   year={2024},
-  publisher={Authorea}
+  DOI = {10.22541/essoar.170965007.78479393/v1}
 }
 
 @article{brinkerhoff2021constraining,
@@ -36,7 +36,7 @@
   number={263},
   pages={385--403},
   year={2021},
-  publisher={Cambridge University Press}
+  DOI={10.1017/jog.2020.112}
 }
 
 @article{he2023hybrid,
@@ -46,7 +46,7 @@
   volume={492},
   pages={112428},
   year={2023},
-  publisher={Elsevier}
+  DOI = {10.1016/j.jcp.2023.112428}
 }
 
 @article{bolibar2023universal,
@@ -56,7 +56,7 @@
   volume={2023},
   pages={1--26},
   year={2023},
-  publisher={G{\"o}ttingen, Germany}
+  DOI = {10.5194/gmd-16-6671-2023}
 }
 
 @article{jouvet2022deep,
@@ -67,7 +67,7 @@
   number={270},
   pages={651--664},
   year={2022},
-  publisher={Cambridge University Press}
+  DOI = {10.1017/jog.2021.120}
 }
 
 @article{jouvet2023inversion,
