+++ paper.bib
@@ -78,7 +78,7 @@
   number={273},
   pages={13--26},
   year={2023},
-  publisher={Cambridge University Press}
+  DOI = {10.1017/jog.2022.41}
 }
 
 @article{jouvet2023ice,
@@ -87,7 +87,7 @@
   journal={Journal of Glaciology},
   pages={1--15},
   year={2023},
-  publisher={Cambridge University Press}
+  DOI = {10.1017/jog.2023.73}
 }
 
 @article{wang2024multi,
@@ -96,13 +96,14 @@
   journal={Journal of Computational Physics},
   pages={112865},
   year={2024},
-  publisher={Elsevier}
+  DOI = {10.1016/j.jcp.2024.112865}
 }
 
 @article{wang2022discovering,
   title={Discovering the rheology of Antarctic Ice Shelves via physics-informed deep learning},
   author={Wang, Yongji and Lai, Ching-Yao and Cowen-Breen, Charlie},
-  journal = {doi.org/10.21203/rs.3.rs-2135795/v1},
+  journal = {SSRN Prepreint},
+  DOI = {10.21203/rs.3.rs-2135795/v1}
   year={2022}
 }
 
@@ -115,12 +116,13 @@
 }
 
 @article{MacAyeal1989,
-	author = {MacAyeal, Douglas R.},
-	title = {Large-scale ice flow over a viscous basal sediment: Theory and application to ice stream B, Antarctica},
-	journal = {Journal of Geophysical Research: Solid Earth},
-	volume = {94},
-	pages = {4071-4087},
-	year = {1989}
+  author = {MacAyeal, Douglas R.},
+  title = {Large-scale ice flow over a viscous basal sediment: Theory and application to ice stream B, Antarctica},
+  journal = {Journal of Geophysical Research: Solid Earth},
+  volume = {94},
+  pages = {4071-4087},
+  year = {1989},
+  DOI = {10.1029/JB094iB04p04071}
 }
 
 
@@ -131,7 +133,7 @@
 	volume={39},
 	pages={91--98},
 	year={1993},
-	publisher={Cambridge University Press}
+	DOI= {10.3189/S0022143000015744}
 }
 
 
@@ -154,19 +156,9 @@
 	number={23},
 	pages={e2309788121},
 	year={2024},
-	publisher={National Acad Sciences}
+	DOI = {10.1073/pnas.2309788121}
 }
 
-
-@article{wang2022thermal,
-	title={Thermal structure of the Amery Ice Shelf from borehole observations and simulations},
-	author={Wang, Yu and Zhao, Chen and Gladstone, Rupert and Galton-Fenzi, Ben and Warner, Roland},
-	journal={The Cryosphere},
-	volume={16},
-	pages={1221--1245},
-	year={2022},
-	publisher={Copernicus GmbH}
-}
 
 
 @article{larour2005rheology,
@@ -198,7 +190,7 @@
   number={3},
   pages={1746--1753},
   year={2013},
-  publisher={Wiley Online Library}
+  DOI = {10.1002/jgrf.20125}
 }
 
 @article{goldberg2011data,
@@ -209,7 +201,7 @@
   number={2},
   pages={315--327},
   year={2011},
-  publisher={Copernicus GmbH}
+  DOI = {10.5194/tc-5-315-2011}
 }
 
 
@@ -220,7 +212,7 @@
   volume={36},
   pages={197--204},
   year={2003},
-  publisher={Cambridge University Press}
+  DOI = {10.3189/172756403781816338}
 }
 
 
@@ -680,13 +672,13 @@
 
 
 @article{morlighem2010spatial,
[…]
