+++ paper.bib
@@ -178,7 +178,7 @@
 	volume={9},
 	pages={1427--1443},
 	year={2015},
-	publisher={Copernicus GmbH}
+	DOI = {10.5194/tc-9-1427-2015}
 }
 
 
@@ -552,7 +552,7 @@
   number={277},
   pages={1167--1186},
   year={2023},
-  publisher={Cambridge University Press}
+  DOI = {10.1017/jog.2023.8}
 }
 
 @article{van2022optimally,
@@ -746,7 +746,7 @@
   volume={492},
   pages={112435},
   year={2023},
-  publisher={Elsevier}
+  DOI = {10.1016/j.jcp.2023.112435}
 }
 
 @article{nye1951flow,
