+++ paper.bib
@@ -103,8 +103,8 @@
   title={Discovering the rheology of Antarctic Ice Shelves via physics-informed deep learning},
   author={Wang, Yongji and Lai, Ching-Yao and Cowen-Breen, Charlie},
   journal = {SSRN Prepreint},
+  year={2022},
   DOI = {10.21203/rs.3.rs-2135795/v1}
-  year={2022}
 }
 
 
@@ -535,13 +535,13 @@
 
 
 @article{riel2021data,
-	title={Data-Driven Inference of the Mechanics of Slip Along Glacier Beds Using Physics-Informed Neural Networks: Case Study on Rutford Ice Stream, Antarctica},
+	title={Data-Driven Inference of the Mechanics of Slip Along Glacier Beds Using Physics-Informed Neural Networks: Case Study on {R}utford {I}ce {S}tream, {A}ntarctica},
 	author={Riel, Bryan and Minchew, Brent and Bischoff, Tobias},
 	journal={Journal of Advances in Modeling Earth Systems},
 	volume={13},
 	pages={e2021MS002621},
 	year={2021},
-	publisher={Wiley Online Library}
+	DOI = {10.1029/2021MS002621}
 }
 
 @article{riel2023variational,
