+++ paper.bib
@@ -484,11 +484,15 @@
 	year = {2021}
 }
 
-@article{daw2022mitigating,
-	title={Mitigating propagation failures in physics-informed neural networks using retain-resample-release (r3) sampling},
-	author={Daw, Arka and Bu, Jie and Wang, Sifan and Perdikaris, Paris and Karpatne, Anuj},
-	journal={arXiv preprint arXiv:2207.02338},
-	year={2022}
+
+@inproceedings{daw2022mitigating,
+  title={Mitigating propagation failures in physics-informed neural networks using retain-resample-release (r3) sampling},
+  author={Daw, Arka and Bu, Jie and Wang, Sifan and Perdikaris, Paris and Karpatne, Anuj},
+  booktitle = {Proceedings of the 40th International Conference on Machine Learning},
+  articleno = {288},
+  numpages = {39},
+  year={2023},
+  DOI = {10.5555/3618408.3618696}
 }
 
 
@@ -514,11 +518,12 @@
 }
 
 
-@article{Charlie2024euler,
-	author = {C. Cown-Breen and Y. Wang and S. Bates and C.-Y. Lai},
-	journal = {Under reivew},
-	title = {Euler operators for mis-specified physics-informed neural networks},
-	year = {2024}
+@inproceedings{charlie2024euler,
+  title={Euler operators for mis-specified physics-informed neural networks},
+  author={Charlie Cowen-Breen and Yongji Wang and Stephen Bates and Ching-Yao Lai},
+  booktitle={ICML 2024 AI for Science Workshop},
+  year={2024},
+  url={https://openreview.net/forum?id=kkGR5fNq2J}
 }
 
 
