+++ paper.bib
@@ -452,7 +452,6 @@
 	publisher={Cambridge University Press}
 }
 
-
 @article{lu2021deepxde,
 	title={DeepXDE: A deep learning library for solving differential equations},
 	author={Lu, Lu and Meng, Xuhui and Mao, Zhiping and Karniadakis, George Em},
@@ -461,7 +460,7 @@
 	number={1},
 	pages={208--228},
 	year={2021},
-	publisher={SIAM}
+	doi = {10.1137/19M1274067}
 }
 
 @article{hornik1989,
@@ -1241,7 +1240,7 @@
   volume={28},
   number={5},
   year={2020},
-  publisher={Brown Univ., Providence, RI (United States)}
+  DOI = {10.4208/cicp.oa-2020-0164}
 }
 
 @article{budd1989review,
