### 2024-08-05 e55f9a6
Update paper.bib
