+++ paper.bib
@@ -517,7 +517,7 @@
 }
 
 
-@inproceedings{charlie2024euler,
+@inproceedings{Charlie2024euler,
   title={Euler operators for mis-specified physics-informed neural networks},
   author={Charlie Cowen-Breen and Yongji Wang and Stephen Bates and Ching-Yao Lai},
   booktitle={ICML 2024 AI for Science Workshop},
