+++ paper.bib
@@ -495,7 +495,7 @@
 }
 
 
-@article{Raissi2020,
+@article{raissi2020,
 	author = {Maziar Raissi and Alireza Yazdani and George Em Karniadakis},
 	journal = {Science},
 	number = {6481},
@@ -506,7 +506,7 @@
 }
 
 
-@article{Raissi2019,
+@article{raissi2019,
 	author = {M. Raissi and P. Perdikaris and G.E. Karniadakis},
 	journal = {Journal of Computational Physics},
 	pages = {686--707},
@@ -517,7 +517,7 @@
 }
 
 
-@inproceedings{Charlie2024euler,
+@inproceedings{charlie2024euler,
   title={Euler operators for mis-specified physics-informed neural networks},
   author={Charlie Cowen-Breen and Yongji Wang and Stephen Bates and Ching-Yao Lai},
   booktitle={ICML 2024 AI for Science Workshop},
