### 2024-08-05 1b9854d
Update paper.bib
