+++ paper.bib
@@ -697,52 +697,6 @@
 }
 
 
-@article{wang2022controls,
-	title={Controls on Larsen C Ice Shelf Retreat From a 60-Year Satellite Data Record},
-	author={Wang, Shujie and Liu, Hongxing and Jezek, Kenneth and Alley, Richard B and Wang, Lei and Alexander, Patrick and Huang, Yan},
-	journal={Journal of Geophysical Research: Earth Surface},
-	volume={127},
-	number={3},
-	pages={e2021JF006346},
-	year={2022},
-	publisher={Wiley Online Library}
-}
-
-@article{weertman1957deformation,
-  title={Deformation of floating ice shelves},
-  author={Weertman, J},
-  journal={Journal of glaciology},
-  volume={3},
-  number={21},
-  pages={38--42},
-  year={1957},
-  publisher={Cambridge University Press}
-}
-@article{sun2020surrogate,
-  title={Surrogate modeling for fluid flows based on physics-constrained deep learning without simulation data},
-  author={Sun, L. and Gao, H. and Pan, S. and Wang, J.X.},
-  journal={Computer Methods in Applied Mechanics and Engineering},
-  volume={361},
-  pages={112732},
-  year={2020},
-  publisher={Elsevier}
-}
-@article{cai2021flow,
-  title={Flow over an espresso cup: inferring 3-D velocity and pressure fields from tomographic background oriented Schlieren via physics-informed neural networks},
-  author={Cai, S. and Wang, Z. and Fuest, F. and Jeon, Y.J. and Gray, C. and Karniadakis, G.E.},
-  journal={Journal of Fluid Mechanics},
-  volume={915},
-  year={2021},
-  publisher={Cambridge University Press}
-}
-@article{cai2022physics,
-  title={Physics-informed neural networks (PINNs) for fluid mechanics: A review},
-  author={Cai, S. and Mao, Z. and Wang, Z. and Yin, M. and Karniadakis, G.E.},
-  journal={Acta Mechanica Sinica},
-  pages={1--12},
-  year={2022},
-  publisher={Springer}
-}
 @article{iwasaki2023one,
   title={One-dimensional ice shelf hardness inversion: Clustering behavior and collocation resampling in physics-informed neural networks},
   author={Iwasaki, Yunona and Lai, Ching-Yao},
@@ -751,485 +705,6 @@
   pages={112435},
   year={2023},
   DOI = {10.1016/j.jcp.2023.112435}
-}
-
-@article{nye1951flow,
-  title={The flow of glaciers and ice-sheets as a problem in plasticity},
-  author={Nye, John F},
-  journal={Proceedings of the Royal Society of London. Series A. Mathematical and Physical Sciences},
-  volume={207},
-  number={1091},
-  pages={554--572},
-  year={1951},
-  publisher={The Royal Society London}
-}
-
-@article{leb1981flow,
-  title={Flow law for polycrystalline ice in glaciers: comparison of theoretical predictions, laboratory data, and field measurements},
-  author={LeB. Hooke, Roger},
-  journal={Reviews of Geophysics},
-  volume={19},
-  number={4},
-  pages={664--672},
-  year={1981},
-  publisher={Wiley Online Library}
-}
-
-@article{glen1952experiments,
-  title={Experiments on the deformation of ice},
-  author={Glen, JW},
-  journal={Journal of Glaciology},
-  volume={2},
-  number={12},
-  pages={111--114},
-  year={1952},
-  publisher={Cambridge University Press}
-}
-
-@book{schulson2009creep,
-  title={Creep and fracture of ice},
-  author={Schulson, Erland M and Duval, Paul},
-  year={2009},
[…]
