### 2024-08-05 d25a9fc
Update paper.bib
