+++ paper.bib
@@ -159,18 +159,6 @@
 	DOI = {10.1073/pnas.2309788121}
 }
 
-
-
-@article{larour2005rheology,
-	title={Rheology of the Ronne Ice Shelf, Antarctica, inferred from satellite radar interferometry data using an inverse control method},
-	author={Larour, Eric and Rignot, E and Joughin, I and Aubry, Denis},
-	journal={Geophysical Research Letters},
-	volume={32},
-	year={2005},
-	publisher={Wiley Online Library}
-}
-
-
 @article{furst2015assimilation,
 	title={Assimilation of Antarctic velocity observations provides evidence for uncharted pinning points},
 	author={F{\"u}rst, JJ and Durand, G and Gillet-Chaulet, F and Merino, N and Tavard, L and Mouginot, J and Gourmelen, N and Gagliardini, O},
@@ -216,95 +204,6 @@
 }
 
 
-@article{glen1955creep,
-	title={The creep of polycrystalline ice},
-	author={Glen, John W},
-	journal={Proceedings of the Royal Society of London. Series A. Mathematical and Physical Sciences},
-	volume={228},
-	number={1175},
-	pages={519--538},
-	year={1955},
-	publisher={The Royal Society London}
-}
-
-
-@article{goldsby2001superplastic,
-	title={Superplastic deformation of ice: Experimental observations},
-	author={Goldsby, DL and Kohlstedt, David L},
-	journal={Journal of Geophysical Research: Solid Earth},
-	volume={106},
-	pages={11017--11030},
-	year={2001},
-	publisher={Wiley Online Library}
-}
-
-
-@article{jezek1985rheology,
-	title={Rheology of glacier ice},
-	author={Jezek, Kenneth C and Alley, Richard B and Thomas, Robert H},
-	journal={Science},
-	volume={227},
-	pages={1335--1337},
-	year={1985},
-	publisher={American Association for the Advancement of Science}
-}
-
-
-@article{thomas1973creep,
-	title={The creep of ice shelves: interpretation of observed behaviour},
-	author={Thomas, Robert H},
-	journal={Journal of Glaciology},
-	volume={12},
-	pages={55--70},
-	year={1973},
-	publisher={Cambridge University Press}
-}
-
-
-@article{thomas1973theory,
-	title={The creep of ice shelves theory},
-	author={Thomas, Robert H},
-	journal={Journal of Glaciology},
-	volume={12},
-	number={64},
-	pages={45--53},
-	year={1973},
-	publisher={Cambridge University Press}
-}
-
-
-@article{schoof2007marine,
-	title={Marine ice-sheet dynamics. Part 1. The case of rapid sliding},
-	author={Schoof, Christian},
-	journal={Journal of Fluid Mechanics},
-	volume={573},
-	pages={27--55},
-	year={2007},
-	publisher={Cambridge University Press}
-}
-
-@article{pegler2018marine,
-	title={Marine ice sheet dynamics: the impacts of ice-shelf buttressing},
-	author={Pegler, Samuel S},
-	journal={Journal of Fluid Mechanics},
-	volume={857},
-	pages={605--647},
-	year={2018},
-	publisher={Cambridge University Press}
-}
-
-
-@article{gudmundsson2013ice,
-	title={Ice-shelf buttressing and the stability of marine ice sheets},
-	author={Gudmundsson, GH},
-	journal={The Cryosphere},
-	volume={7},
-	pages={647--655},
-	year={2013},
-	publisher={Copernicus GmbH}
-}
-
-
 @article{lai2020vulnerability,
 	title={Vulnerability of Antarctica’s ice shelves to meltwater-driven fracture},
[…]
