+++ paper.bib
@@ -265,14 +265,12 @@
 }
 
 
-@inproceedings{daw2022mitigating,
+@article{daw2022mitigating,
   title={Mitigating propagation failures in physics-informed neural networks using retain-resample-release (r3) sampling},
   author={Daw, Arka and Bu, Jie and Wang, Sifan and Perdikaris, Paris and Karpatne, Anuj},
-  booktitle = {Proceedings of the 40th International Conference on Machine Learning},
-  articleno = {288},
-  numpages = {39},
-  year={2023},
-  DOI = {10.5555/3618408.3618696}
+  journal={arXiv preprint arXiv:2207.02338},
+  year={2022},
+  DOI = {10.48550/arXiv.2207.02338}
 }
 
 @article{raissi2019,
