+++ paper.bib
@@ -204,17 +204,6 @@
 }
 
 
-@article{lai2020vulnerability,
-	title={Vulnerability of Antarctica’s ice shelves to meltwater-driven fracture},
-	author={Lai, Ching-Yao and Kingslake, Jonathan and Wearing, Martin G and Chen, Po-Hsuan Cameron and Gentine, Pierre and Li, Harold and Spergel, Julian J and van Wessem, J Melchior},
-	journal={Nature},
-	volume={584},
-	pages={574--578},
-	year={2020},
-	publisher={Nature Publishing Group}
-}
-
-
 @misc{Mouginot2019velo, 
 	title={MEaSUREs Phase-Based Antarctica Ice Velocity Map, Version 1}, 
 	url={https://nsidc.org/data/NSIDC-0754/versions/1}, 
@@ -232,16 +221,6 @@
 	year={2020} 
 }
 
-@article{morlighem2020deep,
-	title={Deep glacial troughs and stabilizing ridges unveiled beneath the margins of the Antarctic ice sheet},
-	author={Morlighem, Mathieu and Rignot, Eric and Binder, Tobias and Blankenship, Donald and Drews, Reinhard and Eagles, Graeme and Eisen, Olaf and Ferraccioli, Fausto and Forsberg, Ren{\'e} and Fretwell, Peter and others},
-	journal={Nature Geoscience},
-	volume={13},
-	number={2},
-	pages={132--137},
-	year={2020},
-	publisher={Nature Publishing Group}
-}
 
 @article{lu2021deepxde,
 	title={DeepXDE: A deep learning library for solving differential equations},
@@ -253,17 +232,6 @@
 	year={2021},
 	doi = {10.1137/19M1274067}
 }
-
-
-@article{Karniadakis2021,
-	author = {George Em Karniadakis and Ioannis G. Kevrekidis and Lu Lu and Paris Perdikaris and Sifan Wang and Liu Yang},
-	journal = {Nature Reviews Physics},
-	pages = {422--440},
-	title = {Physics-informed machine learning},
-	volume = {3},
-	year = {2021}
-}
-
 
 @article{daw2022mitigating,
   title={Mitigating propagation failures in physics-informed neural networks using retain-resample-release (r3) sampling},
