### 2024-08-05 826a3cd
Delete equation/README.md
