### 2024-08-05 0899063
Delete model/README.md
