+++ tutorial/README.md
@@ -25,7 +25,7 @@ $$\begin{array}{l}
 $$ \qquad \nabla \cdot (hu) = \frac{\partial (hu)} {\partial x} + \frac{\partial (hv)} {\partial y} = \dot{a} - \dot{b} \qquad \qquad \text{(Steady mass conserv.)} $$
 
 where $(u, v)$, $h$ and $\mu$ are the velocity vector, thickness and viscosity of the floating 
-ice, respectively. $\dot{a}$ and $\dot{b}$ are the snow accumulation rate and basel melting rate,
+ice, respectively.  $\dot{a}$ and $\dot{b}$  are the snow accumulation rate and basel melting rate,
 respectively. For the tutorial example provided in this folder, we consider floating 
 ice moving in a confined rectangular channel. For simplicity, we assume that both $\dot{a}$ and $\dot{b}$
 are equal to 0. The domain size and the associated boundary conditions for the ice flow are listed 
