+++ docs/Codes.md
@@ -114,3 +114,17 @@ the remote-sensing velocity data (450m resolution). Although users can modify th
 output on other higher-resolution grids. Furthermore, the prediction function for **XPINNs**
 will automatically stitch the network outputs from different sub-regions into a single large domain that 
 covers all the sub-regions.
+
+
+## Folder: `Optimizer`
+
+The `optimization.py` provides two optimization methods. 
+
+[**Adam**](https://arxiv.org/pdf/1412.6980): a first-order gradient-based optimization method 
+based on adaptive estimates of lower-order moments.
+
+[**L-BFGS**](http://users.iems.northwestern.edu/~nocedal/PDFfiles/limited-memory.pdf): Limited-memory BFGS method, a second-order quasi-Newton optimization method for
+solving unconstrained nonlinear optimization problems, using a limited amount of computer memory.
+
+A **stochastic training scheme** is applied to two optimizers, where the code randomizes both data samples and collocation points at regular interval during training 
+to minimize the cheating effect.
