+++ docs/Codes.md
@@ -7,11 +7,8 @@ including the normalization, flattening and random sampling, and the post-proces
 network output, including re-normalization and reshaping. Two versions of codes are provided in this 
 folder. The `pinns` subfolder involves the code is for the **regular PINN** training,  and the `xpinns` 
 subfolder is for the **extended-PINN (XPINN**) training.
-
- <br />
  
-### `DIFFICE_jax/data/pinns/preprocessing.py`
-### `DIFFICE_jax/data/xpinns/preprocessing.py`
+### `DIFFICE_jax/data/pinns/preprocessing.py` / `DIFFICE_jax/data/xpinns/preprocessing.py`
 
 Involving essential functions to normalize the raw data loaded from the MATLAB format (`.mat`). The 
 function in the script will automatically detect the characterisitc scale of each variable in the 
@@ -19,10 +16,8 @@ dataset, including the spatial coordiates $(x,y)$, velocity $(u, v)$ and thickne
 normalize those variables to be within the range of $[-1, 1]$. The script also re-organizes and reshapes
 the data that is required by the code in the `model` folder to ensure the success of the PINN training.
 
- <br />
  
-### `DIFFICE_jax/data/pinns/sampling.py`
-### `DIFFICE_jax/data/xpinns/sampling.py`
+### `DIFFICE_jax/data/pinns/sampling.py` / `DIFFICE_jax/data/xpinns/sampling.py`
 
 Involving essential functions to sample a batch of pre-processed data used for the PINN training. Users
 can specify the number of sampling points for both velocity and thickness data, as well as the collocation 
