+++ docs/Codes.md
@@ -34,16 +34,12 @@ for each sub-region, not the entire domain.
 
 The folder includes the codes that compute the residues of governing equations and boundary conditions 
 involved in the PINN training. 
-
- <br />
  
 ### `DIFFICE_jax/equation/ssa_eqn_iso.py`
 
 involving functions to compute the residue of the normalized **isotropuc** Shallow-Shelf Approximation (SSA) 
 equations and their associated dynamic boundary conditions at the calving front. Both the SSA equations and
 the boundary conditions are given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/paper.md) .
-
- <br />
  
 ### `DIFFICE_jax/equation/ssa_eqn_aniso_zz.py`
 
@@ -73,10 +69,8 @@ component for PINN training. Detailed instructions for correctly calling the fun
 scripts can be found in the [example codes](https://github.com/YaoGroup/DIFFICE_jax/tree/main/examples) 
 within the `examples` folder of this repository.
 
- <br />
 
-### `DIFFICE_jax/model/pinns/initialization.py`
-### `DIFFICE_jax/model/xpinns/initialization.py`
+### `DIFFICE_jax/model/pinns/initialization.py` / `DIFFICE_jax/model/xpinns/initialization.py`
 
 Involving essential functions to intialize weights and biases for all neural networks required 
 for the problem. The code use Xavier initialization scheme, so that the weights between each 
@@ -84,10 +78,8 @@ two layers are generated following a truncated normal distribution with zero mea
 variance equal to $2/(n_{l-1}+n_{l})$, where $n_{l}$ and $n_{l+1}$ indicates the number of
 units in the previous and next layers. The biases are initialized with all zero.
 
- <br />
- 
-### `DIFFICE_jax/model/pinns/networks.py`
-### `DIFFICE_jax/model/xpinns/networks.py`
+
+### `DIFFICE_jax/model/pinns/networks.py` / `DIFFICE_jax/model/xpinns/networks.py`
 
 Involving essential functions to generate the neural network model for each physical variable 
 involved in the problem. For regular PINN training, two networks are created. One network has 
@@ -98,10 +90,8 @@ fully-connected multiple-layer preceptrons (MLP) using `tanh` as the default act
 When creating the neural networks, users need to specify whether the neural networks are created for isotropic 
 and anistropic viscosity, as these two cases requires different number of outputs.
 
- <br />
  
-### `DIFFICE_jax/model/pinns/loss.py`
-### `DIFFICE_jax/model/xpinns/loss.py`
+### `DIFFICE_jax/model/pinns/loss.py` / `DIFFICE_jax/model/xpinns/loss.py`
 
 Involving essential functions to generate the total loss function for PINN training on assimilating
 remote-sensing data of ice shelves and inferring their effective viscosity. The mathematical formation
@@ -111,10 +101,8 @@ The loss function for the XPINN training is given [here](https://github.com/YaoG
 We note that users should call different functions in the `loss.py` script to generate the loss functions
 for inferring isotropic (i.e. `loss_iso_create()`) and anisotropic viscosity (i.e. `loss_aniso_create()`)
 
- <br />
  
-### `DIFFICE_jax/model/pinns/prediction.py`
-### `DIFFICE_jax/model/xpinns/prediction.py`
+### `DIFFICE_jax/model/pinns/prediction.py` / `DIFFICE_jax/model/xpinns/prediction.py`
 
 Involving functions to predict the neural network output at the high-resolution grids for evaluation or 
 visualization. The default setting of the function is to predict the data at the same resolution grid of
