### 2024-08-05 4128784
Delete optimizer/README.md
