### 2024-08-05 ec743c7
Delete src/old version/version 0.0 directory
