### 2024-08-05 df01391
Update XPINNs.md
