+++ README.md
@@ -5,7 +5,7 @@ A user-friendly DIFFerentiable neural-network solver for data assimilation of IC
 `DIFFICE_jax` is a Python package that solves the depth-integrated Stokes equation for `ice shelves`, and can be adopted for `ice sheets` by modifying the partial differential equations (PDE) in the neural network loss function. It uses PDEs to interpolate descretized remote-sensing data into meshless and differentible functions, and infer ice shelves' viscosity structure via back-propagation and automatic differentiation (AD). The algorithm is based on physics-informed neural networks (`PINNs`) [@Raissi2019] and implemented in JAX [@jax2018github]. The `DIFFICE_jax` package involves several advanced features in addition to vanilla PINNs algorithms, including collocation points resampling, non-dimensionalization of the data adnd equations, extended PINN (see figure below and [XPINN documentation](https://github.com/YaoGroup/DIFFICE_jax/blob/main/model/XPINNs.md)), viscosity exponential scaling function, which are essential for accurate inversion. The package is designed to be user-friendly and accessible for beginners. The Github respository also provides tutorial examples for users at different levels to help master the method.
 
 <p align="center">
-    <img src="model/xpinns.png" alt="results" width="90%">
+    <img src="docs/figure/xpinns.png" alt="results" width="90%">
 </p>
 
 ## Installation
