+++ examples/README.md
@@ -5,9 +5,9 @@ and thickness data and infer effective ice viscosity under either **isotropic** 
 via **regular PINNs** or **extended-PINNs (XPINNs)**. 
 
 The mathematical formulation for inferring **isotropic** ice viscosity
-via regular PINNs are provided in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/paper.md).  The 
-description for inferring **anisotropic** viscosity is given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/examples/Anisotropic.md),
-and the description of **XPINNs** settings is given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/model/XPINNs.md).
+via regular PINNs are provided in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/Isotropic.md).  The 
+description for inferring **anisotropic** viscosity is given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/Anisotropic.md),
+and the description of **XPINNs** settings is given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/model/docs/XPINNs.md).
 
 
 ## `examples/train_pinns_iso.py`
