### 2024-08-05 3c46291
Update Anisotropic.md
