### 2024-08-05 583cdcf
Update Anisotropic.md
