+++ docs/XPINNs.md
@@ -87,4 +87,4 @@ where $\mathcal{L_d}$ and $\mathcal{L_e}$ are the data loss and equation loss as
 $\gamma_e$ and $\gamma_b$ are the weighting pre-factors for the equation and boundary condition loss. An extra
 regularization loss term $\mathcal{L}_{reg}$ is required to infer anisotropic viscosity. The principle of this loss
 terms and the appropriate value of its weight $\gamma_g$ are detailed described in 
-[this page](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/Anisotropic.md).
+[here](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/Anisotropic.md).
