+++ docs/Data_prepare.md
@@ -46,7 +46,7 @@ all of them should be saved in a `cell` type. Additionally, we note that all the
 `ud` and `vd`) should have the exact same shape, and the same applies to the thickness-related quantities (`xd_h`, `yd_h`
 and `h_d`). The calving front-related quantities (`xct`, `yct` and `nnct`) should also have the same length. 
 
-Apart from the above quantities, below are the additional quantities particularly required for the **XPINN** training
+Apart from the above quantities, below are the additional quantities particularly required for the **XPINNs** training
 
 |Variables | meaning | XPINNs|
 | ------------- | ------------- | ------------ |
@@ -60,7 +60,7 @@ Apart from the above quantities, below are the additional quantities particularl
 | `idxcrop_h` | vertex position of sub-regional thickness data in the whole domain | cell[4x1 array]|
 
 `x_md` and `y_md` are the positions of the interface between each two consecutive sub-regions, which are required to 
-evaluate the [continuity loss](https://github.com/YaoGroup/DIFFICE_jax/blob/main/model/XPINNs.md) for the **XPINN**
+evaluate the [continuity loss](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/XPINNs.md) for the **XPINNs**
 training. The other 6 quantities are the required information to merge the variable of all sub-regions into a 
 large entity matrix, representing the whole domain.
 
