+++ docs/Data.md
@@ -1,11 +1,16 @@
 # Data source
 
-This folder contains several examples of remote-sensing velocity and thickness data for different 
-ice shelves. The dataset for each ice shelf is directly cropped from the raw data files published online, 
-maintaining the same resolution. No processing has been conducted on these data. The raw velocity data is downloaded
-from NASA MEaSUREs Phase-Based Antarctica Ice Velocity Map, Version 1 [(NSIDC-0754)](https://nsidc.org/data/nsidc-0754/versions/1) 
-with a 450m resolution. The thickness data is downloaded from NASA MEaSUREs BedMachine Antarctica, Version 2
-[(NSIDC-0756)](https://nsidc.org/data/nsidc-0756/versions/2) with a 500m resolution.
+The remote-sensing data of both ice-shelf velocity and thickness that can be used for the `DIFFICE_jax` package 
+are available online The raw velocity data can be downloaded from NASA MEaSUREs Phase-Based Antarctica Ice Velocity Map, 
+Version 1 [(NSIDC-0754)](https://nsidc.org/data/nsidc-0754/versions/1) with a 450m resolution. The thickness data can be
+downloaded from NASA MEaSUREs BedMachine Antarctica, Version 2 [(NSIDC-0756)](https://nsidc.org/data/nsidc-0756/versions/2)
+with a 500m resolution. These raw datasets are not provided in the package. Instead, the datasets provided in the `examples`
+folder are cropped from these raw datasets under the same resolution with no extra processing, and saved separately for each
+ice shelf. Additional information that is required for the PINN training,  such as the position of the ice-shelf calving front,
+is also included in the dataset. The full requirement of the dataset that can ensure the success of PINN training is provided below. 
+Users should strictly follow these requirements in order to create datasets for ice shelves that are not currently available 
+in the package.
+
 
  <br />
  
