+++ docs/Data.md
@@ -1,16 +1,15 @@
 # Data source
 
-The remote-sensing data of both ice-shelf velocity and thickness that can be used for the `DIFFICE_jax` package 
-are available online The raw velocity data can be downloaded from NASA MEaSUREs Phase-Based Antarctica Ice Velocity Map, 
+The remote-sensing data of both ice-shelf velocity and thickness used for the `DIFFICE_jax` package 
+are available online. The raw velocity data can be downloaded from NASA MEaSUREs Phase-Based Antarctica Ice Velocity Map, 
 Version 1 [(NSIDC-0754)](https://nsidc.org/data/nsidc-0754/versions/1) with a 450m resolution. The thickness data can be
 downloaded from NASA MEaSUREs BedMachine Antarctica, Version 2 [(NSIDC-0756)](https://nsidc.org/data/nsidc-0756/versions/2)
 with a 500m resolution. These raw datasets are not provided in the package. Instead, the datasets provided in the `examples`
-folder are cropped from these raw datasets under the same resolution with no extra processing, and saved separately for each
-ice shelf. Additional information that is required for the PINN training,  such as the position of the ice-shelf calving front,
-is also included in the dataset. The full requirement of the dataset that can ensure the success of PINN training is provided below. 
-Users should strictly follow these requirements in order to create datasets for ice shelves that are not currently available 
-in the package.
-
+folder are cropped from these raw datasets at the same resolution with no additional processing, and are saved separately for each
+ice shelf. Additional information required for the PINN training,  such as the position of the ice-shelf calving front,
+is also included in the dataset. The full requirements of the dataset, which ensures the success of PINN training, are provided
+in the document below.  Users should strictly follow these requirements to create datasets for ice shelves that are not currently
+available in the package.
 
  <br />
  
