+++ examples/README.md
@@ -10,7 +10,7 @@ description for inferring **anisotropic** viscosity is given in [this link](http
 and the description of **XPINNs** settings is given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/XPINNs.md).
 
 
-## `examples/train_pinns_iso.py`
+### `examples/train_pinns_iso.py`
 
 A python script that assimilate remote-sensing data and infer the effective ice viscosity under **isotropic**
 assumption via **regular PINNs**. The code are computationally-efficient and accurate enough to study ice shelves
@@ -19,7 +19,7 @@ provided in the `colab` subfolder. View it by clicking
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/YaoGroup/DIFFICE_jax/blob/main/examples/colab/train_pinns_iso.ipynb)
 
 
-## `examples/train_pinns_aniso.py`
+### `examples/train_pinns_aniso.py`
 
 A python script that assimilate remote-sensing data and infer the effective ice viscosity under **anisotropic**
 assumption via **regular PINNs**. Different from isotropic viscosity inversion, this code infers two viscosity 
@@ -28,7 +28,7 @@ provided in the `colab` subfolder. View it by clicking
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/YaoGroup/DIFFICE_jax/blob/main/examples/colab/train_pinns_aniso.ipynb)
 
 
-## `examples/train_xpinns_iso.py`
+### `examples/train_xpinns_iso.py`
 
 A python script that assimilate remote-sensing data and infer the effective ice viscosity under **isotropic**
 assumption via **extended-PINNs (XPINNs)**. This code are required to study several largest ice shelves around the
@@ -38,7 +38,7 @@ An companion Colab Notebook of this script is provided in the `colab` subfolder.
 [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/YaoGroup/DIFFICE_jax/blob/main/examples/colab/train_xpinns_iso.ipynb)
 
 
-## `examples/train_xpinns_aniso.py`
+### `examples/train_xpinns_aniso.py`
 
 A python script that assimilate remote-sensing data and infer the effective ice viscosity under **anisotropic**
 assumption via **extended-PINNs (XPINNs)**. Different from isotropic viscosity inversion, this code infers two viscosity 
