### 2024-08-05 522d46b
Update train_syndata.py
