### 2024-08-05 e8f6247
Update train_pinns_aniso.py
