### 2024-08-05 514e135
Update train_pinns_iso.py
