+++ examples/train_pinns_iso.py
@@ -57,7 +57,7 @@ shelfname = 'Amery'
 
 # create the dataset filename
 filename = 'data_pinns_' + shelfname + '.mat'
-filepath = str(project_root.joinpath('data').joinpath(filename))
+filepath = str(rootdir.joinpath('real_data').joinpath(filename))
 
 # create the output file name
 outputName = shelfname + f'_pinns_iso_seed={seed:.0f}'
