### 2024-08-05 f5474fc
Update train_xpinns_aniso.py
