### 2024-08-05 fe953a5
Update train_xpinns_iso.py
