+++ diffice_jax/__init__.py
@@ -1,8 +1,26 @@
-from .data.pinns.preprocessing import normalize_data as normdata
-from .data.xpinns.preprocessing import normalize_data as normdatax
+from .data.pinns.preprocessing import normalize_data as normdata_pinn
+from .data.xpinns.preprocessing import normalize_data as normdata_xpinn
 
+from .data.pinns.sampling import data_sample_create as dsample_pinn
+from .data.xpinns.sampling import data_sample_create as dsample_xpinn
 
-from .data.pinns.sampling import data_sample_create as dsample
-from .data.xpinns.sampling import data_sample_create as dsamplex
+from .equation import ssa_eqn_iso
+from .equation import ssa_eqn_aniso_zz
 
+from .model.pinns.networks import solu_create as solu_pinn
+from .model.xpinns.networks import solu_create as solu_xpinn
 
+from .model.pinns.loss import loss_iso_create as loss_iso_pinn
+from .model.pinns.loss import loss_aniso_create as loss_aniso_pinn
+from .model.xpinns.loss import loss_iso_create as loss_iso_xpinn
+from .model.xpinns.loss import loss_aniso_create as loss_aniso_xpinn
+
+from .model.pinns.prediction import predict as predict_pinn
+from .model.xpinns.prediction import predict as predict_xpinn
+
+from .optimizer.optimization import adam_optimizer, lbfgs_optimizer
+
+__all__ = ["normdata_pinn", "normdata_xpinn", "dsample_pinn", "dsample_xpinn",
+           "ssa_eqn_iso", "ssa_eqn_aniso", "solu_pinn", "solu_xpinn", 
+           "loss_iso_pinn", "loss_aniso_pinn", "loss_iso_xpinn", "loss_aniso_xpinn",
+           "predict_pinn", "predict_xpinn", "adam_optimizer", "lbfgs_optimizer"]
