### 2024-08-06 e0bc1fe
Update pyproject.toml
