### 2024-08-06 0f5f5c1
Update pyproject.toml
