+++ pyproject.toml
@@ -8,7 +8,7 @@ include-package-data = true
 
 [project]
 name = "DIFFICE_jax"
-version = "0.0.3"
+version = "0.0.4"
 authors = [
   { name="Yongji Wang", email="yw8211@nyu.edu" },
 ]
