+++ pyproject.toml
@@ -3,8 +3,11 @@ requires = ["setuptools", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["diffice_jax"]
+packages = {find = {}}
 include-package-data = true
+
+[tool.setuptools.package-data]
+"my_package" = ["*.py"]
 
 [project]
 name = "DIFFICE_jax"
