+++ pyproject.toml
@@ -11,7 +11,7 @@ include-package-data = true
 
 [project]
 name = "DIFFICE_jax"
-version = "0.0.4"
+version = "0.0.5"
 authors = [
   { name="Yongji Wang", email="yw8211@nyu.edu" },
 ]
