### 2024-08-06 a4dd03c
Update ssa_eqn_aniso_zz.py
