### 2024-08-06 e073ce8
Update and rename ssa_eqn_aniso_zz.py to eqn_aniso_zz.py
