### 2024-08-06 34af4ae
Update and rename ssa_eqn_iso.py to eqn_iso.py
