+++ diffice_jax/equation/eqn_iso.py
@@ -43,6 +43,8 @@ def vectgrad(func, x):
     return grad_all, sol
 
 
+#%% Isotropic shallow-shelf approximation (SSA) equations in the normalized form
+
 def gov_eqn(net, x, scale):
     """
     :param net: the neural net instance for calculating the informed part
@@ -62,8 +64,6 @@ def gov_eqn(net, x, scale):
     rv0 = v0 / u0m
     rx0 = lx0 / l0m
     ry0 = ly0 / l0m
-
-    mu0 = rho * gd * h0 * (l0m / u0m)
 
     def grad1stOrder(net, x):
         grad, sol = vectgrad(net, x)
@@ -127,8 +127,6 @@ def front_eqn(net, x, nb, scale):
     rv0 = v0 / u0m
     rx0 = lx0 / l0m
     ry0 = ly0 / l0m
-
-    mu0 = rho * gd * h0 * (l0m / u0m)
 
     grad, sol = vectgrad(net, x)
     h = sol[:, 2:3]
