### 2024-08-06 72bb5ab
Update eqn_iso.py
