+++ docs/Codes.md
@@ -1,7 +1,7 @@
 # Code description
 
 
-## Folder: `DIFFICE_jax/data` 
+## Folder: `diffice_jax/data` 
 This folder provides the codes that conduct the pre-preprocessing of the input data,
 including the normalization, flattening and random sampling, and the post-processing of the neural
 network output, including re-normalization and reshaping. Two versions of codes are provided in this 
@@ -16,7 +16,7 @@ scripts can be found in the [example codes](https://github.com/YaoGroup/DIFFICE_
 within the `examples` folder of this repository.
 
  
-### `DIFFICE_jax/data/pinns/preprocessing.py` / `DIFFICE_jax/data/xpinns/preprocessing.py`
+### `diffice_jax/data/pinns/preprocessing.py` / `diffice_jax/data/xpinns/preprocessing.py`
 
 Involving essential functions to normalize the raw data loaded from the MATLAB format (`.mat`). The 
 function in the script will automatically detect the characterisitc scale of each variable in the 
@@ -25,7 +25,7 @@ normalize those variables to be within the range of $[-1, 1]$. The script also r
 the data that is required by the code in the `model` folder to ensure the success of the PINN training.
 
  
-### `DIFFICE_jax/data/pinns/sampling.py` / `DIFFICE_jax/data/xpinns/sampling.py`
+### `diffice_jax/data/pinns/sampling.py` / `diffice_jax/data/xpinns/sampling.py`
 
 Involving essential functions to sample a batch of pre-processed data used for the PINN training. Users
 can specify the number of sampling points for both velocity and thickness data, as well as the collocation 
@@ -38,18 +38,18 @@ for each sub-region, not the entire domain.
  <br />
  
 
-## Folder: `DIFFICE_jax/equation` 
+## Folder: `diffice_jax/equation` 
 
 The folder includes the codes that compute the residues of governing equations and boundary conditions 
 involved in the PINN training. 
  
-### `DIFFICE_jax/equation/ssa_eqn_iso.py`
+### `diffice_jax/equation/eqn_iso.py`
 
 involving functions to compute the residue of the normalized **isotropuc** Shallow-Shelf Approximation (SSA) 
 equations and their associated dynamic boundary conditions at the calving front. Both the SSA equations and
 the boundary conditions are given in [this link](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/Isotropic.md) .
  
-### `DIFFICE_jax/equation/ssa_eqn_aniso_zz.py`
+### `diffice_jax/equation/eqn_aniso_zz.py`
 
 involving functions to compute the residue of the normalized **anisotropic** Shallow-Shelf Approximation (SSA) 
 equations and their associated dynamic boundary conditions at the calving front. The suffix `_zz` indicate the
@@ -61,7 +61,7 @@ the associated boundary conditions are given in [this link](https://github.com/Y
 
  
 
-## Folder: `DIFFICE_jax/model`
+## Folder: `diffice_jax/model`
 
 This folder provides the core codes for generating neural network, creating
 loss function and predicting output variables for the PINN training on assimilating 
@@ -71,7 +71,7 @@ the **regular PINN** training, and the `xpinns` folder is for the **extended-PIN
 (XPINN)** training. 
 
 
-### `DIFFICE_jax/model/pinns/initialization.py` / `DIFFICE_jax/model/xpinns/initialization.py`
+### `diffice_jax/model/pinns/initialization.py` / `diffice_jax/model/xpinns/initialization.py`
 
 Involving essential functions to intialize weights and biases for all neural networks required 
 for the problem. The code use Xavier initialization scheme, so that the weights between each 
[…]
