+++ README.md
@@ -20,8 +20,9 @@ python -m pip install DIFFICE_jax
 ## Documentation
 
 The documentation for the algorithms and the mathematical formulation for the data assimilation of ice shelves 
-are provided in the `docs` folder. Documentations for the `tutorial` and real-data `examples` are separately 
-given in each folder.
+are provided in the [`docs`](https://github.com/YaoGroup/DIFFICE_jax/tree/main/docs)  folder. Documentations for the **tutorial** examples and **real-data** examples are  
+given in the  [`tutorial`](https://github.com/YaoGroup/DIFFICE_jax/tree/main/tutorial)  folder and [`examples`](https://github.com/YaoGroup/DIFFICE_jax/tree/main/examples) 
+folders, respectively.
 
 ## Getting start with a Tutorial
 We highly recommend the user who has no previous experience in either PINNs or inverse problems in Glaciology to get familar
