### 2024-08-06 e042113
Update pyproject.toml
