+++ pyproject.toml
@@ -11,7 +11,7 @@ include-package-data = true
 
 [project]
 name = "DIFFICE_jax"
-version = "0.0.5"
+version = "0.0.6"
 authors = [
   { name="Yongji Wang", email="yw8211@nyu.edu" },
 ]
