+++ .github/workflows/python-publish.yml
@@ -36,4 +36,4 @@ jobs:
       uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
       with:
         user: __token__
-        password: ${{ secrets.TESTPYPI_TOKEN }}
+        password: ${{ secrets.PYPI_API_TOKEN }}
