+++ pyproject.toml
@@ -11,7 +11,7 @@ include-package-data = true
 
 [project]
 name = "DIFFICE_jax"
-version = "0.0.1"
+version = "0.0.7"
 authors = [
   { name="Yongji Wang", email="yw8211@nyu.edu" },
 ]
