+++ diffice_jax/data/xpinns/sampling.py
@@ -1,3 +1,7 @@
+"""
+@author: Yongji Wang
+"""
+
 import jax.numpy as jnp
 from jax import random
 from jax.tree_util import tree_map
