+++ pyproject.toml
@@ -11,7 +11,7 @@ include-package-data = true
 
 [project]
 name = "DIFFICE_jax"
-version = "0.0.7"
+version = "0.0.8"
 authors = [
   { name="Yongji Wang", email="yw8211@nyu.edu" },
   { name="Ching-Yao Lai", email="cyaolai@stanford.edu" },
