+++ diffice_jax/model/xpinns/prediction.py
@@ -8,7 +8,7 @@ from pathlib import Path
 project_root = Path(__file__).parent.parent.parent
 sys.path.append(str(project_root))
 
-from equation.ssa_eqn_iso import vectgrad
+from equation.eqn_iso import vectgrad
 
 
 def dataArrange(var, idxval, dsize):
