+++ diffice_jax/model/xpinns/initialization.py
@@ -18,7 +18,7 @@ def init_single_net(parent_key, layer_widths):
 
 
 # generate weights and biases for all networks required in the XPINNs problem
-def init_xpinns(parent_key, n_hl, n_unit, n_sub=1, aniso=False):
+def init_nets(parent_key, n_hl, n_unit, n_sub=1, aniso=False):
     '''
     :param n_hl: number of hidden layers [int]
     :param n_unit: number of units in each layer [int]
