+++ README.md
@@ -31,7 +31,7 @@ that allows users to compare the given viscosity with the PINN inferred viscosit
 
 ## Real-data Examples 
 The inversion codes with the real velocity and thickness data for **four** different ice shelves surrounding the Antarctica are provided in the [`examples`](https://github.com/YaoGroup/DIFFICE_jax/tree/main/examples) folders. In the [paper](https://github.com/YaoGroup/DIFFICE_jax/tree/main/paper.md), we
-summarized **six algorithm features** of the `DIFFICE_jax` package beyond the vanilla PINNs code. We provide four example codes that can be used to analyze different ice-shelf datasets. For each example code, the corresponding implemented features and the ice-shelf dataset it can analyze are listed in the table below. 
+summarized **six algorithm features** of the `DIFFICE_jax` package beyond the vanilla PINNs code. We provide four example codes in [`examples`](https://github.com/YaoGroup/DIFFICE_jax/tree/main/examples) that can be used to analyze different ice-shelf datasets. For each example code, the corresponding implemented features and the ice-shelf dataset it can analyze are listed in the table below. 
 
 | Example codes  | Feature # | Ice shelf |
 | ------------- | ------------- | ------------- |
