### 2024-08-08 fd39e23
Update train_syndata.py
