### 2024-08-08 cda2080
Update train_syndata.py
