### 2024-08-08 2ec10cf
Update train_pinns_iso.py
