+++ examples/train_pinns_iso.py
@@ -8,17 +8,12 @@ from scipy.io import savemat, loadmat
 from pathlib import Path
 import pickle
 
-project_root = Path(__file__).parent.parent
-sys.path.append(str(project_root))
-
-from data.pinns.preprocessing import normalize_data
-from data.pinns.sampling import data_sample_create
-from equation.ssa_eqn_iso import vectgrad, gov_eqn, front_eqn
-from model.pinns.initialization import init_pinns
-from model.pinns.networks import solu_create
-from model.pinns.loss import loss_iso_create
-from model.pinns.prediction import predict
-from optimizer.optimization import adam_optimizer, lbfgs_optimizer
+from diffice_jax import normdata_pinn, dsample_pinn
+from diffice_jax import vectgrad, ssa_iso, dbc_iso
+from diffice_jax import init_pinn, solu_pinn
+from diffice_jax import loss_iso_pinn
+from diffice_jax import predict_pinn
+from diffice_jax import adam_opt, lbfgs_opt
 
 # find the root directory
 rootdir = Path(__file__).parent
@@ -40,10 +35,10 @@ n_unit = 40
 lw = [0.05, 0.1]
 
 # number of sampling points
-n_smp = 7000    # for velocity data
-nh_smp = 6500   # for thickness data
-n_col = 7000    # for collocation points
-n_cbd = 600     # for boundary condition (calving front)
+n_smp = 8000    # for velocity data
+nh_smp = 7500   # for thickness data
+n_col = 8000    # for collocation points
+n_cbd = 800     # for boundary condition (calving front)
 # group all the number of points
 n_pt = jnp.array([n_smp, nh_smp, n_col, n_cbd], dtype='int32')
 # double the points for L-BFGS training
@@ -71,7 +66,7 @@ if not isExist:
 # load the datafile
 rawdata = loadmat(filepath)
 # obtain the data for training
-data_all = normalize_data(rawdata)
+data_all = normdata_pinn(rawdata)
 scale = data_all[4][0:2]
 
 
@@ -81,20 +76,20 @@ scale = data_all[4][0:2]
 trained_params = init_pinns(keys[0], n_hl, n_unit)
 
 # create the solution function
-pred_u = solu_create()
+pred_u = solu_pinn()
 
 # create the data function for Adam
-dataf = data_sample_create(data_all, n_pt)
+dataf = dsample_pinn(data_all, n_pt)
 keys_adam = random.split(keys[1], 5)
 data = dataf(keys_adam[0])
 # create the data function for L-BFGS
-dataf_l = data_sample_create(data_all, n_pt2)
+dataf_l = dsample_pinn(data_all, n_pt2)
 key_lbfgs = keys[2]
 
 # group the gov. eqn and bd cond.
-eqn_all = (gov_eqn, front_eqn)
+eqn_all = (ssa_iso, dbc_iso)
 # calculate the loss function
-NN_loss = loss_iso_create(pred_u, eqn_all, scale, lw)
+NN_loss = loss_iso_pinn(pred_u, eqn_all, scale, lw)
 # calculate the initial loss and set it as the loss reference value
 NN_loss.lref = NN_loss(trained_params, data)[0]
 
@@ -113,12 +108,12 @@ lr = 1e-3
 start_time = time.time()
 
 """training with Adam"""
-trained_params, loss1 = adam_optimizer(keys_adam[0], NN_loss, trained_params, dataf, epoch1, lr=lr)
+trained_params, loss1 = adam_opt(keys_adam[0], NN_loss, trained_params, dataf, epoch1, lr=lr)
 
 # sample the data for L-BFGS training
 data_l = dataf_l(key_lbfgs)
 """training with L-BFGS"""
[…]
