### 2024-08-08 7a2cf08
Update train_syndata.py
