### 2024-08-08 dea1624
Update train_pinns_aniso.py
