### 2024-08-08 1955d2e
Update train_xpinns_iso.py
