### 2024-08-08 0ddefa0
Update train_xpinns_aniso.py
