+++ examples/train_xpinns_aniso.py
@@ -9,17 +9,12 @@ from scipy.io import savemat, loadmat
 from pathlib import Path
 import pickle
 
-project_root = Path(__file__).parent.parent
-sys.path.append(str(project_root))
-
-from data.xpinns.preprocessing import normalize_data
-from data.xpinns.sampling import data_sample_create
-from equation.ssa_eqn_aniso_zz import vectgrad, gov_eqn, front_eqn
-from model.xpinns.initialization import init_xpinns
-from model.xpinns.networks import solu_create
-from model.xpinns.loss import loss_aniso_create
-from model.xpinns.prediction import predict
-from optimizer.optimization import adam_optimizer, lbfgs_optimizer
+from diffice_jax import normdata_xpinn, dsample_xpinn
+from diffice_jax import vectgrad, ssa_aniso, dbc_aniso
+from diffice_jax import init_xpinn, solu_xpinn
+from diffice_jax import loss_aniso_xpinn
+from diffice_jax import predict_xpinn
+from diffice_jax import adam_opt, lbfgs_opt
 
 # find the root directory
 rootdir = Path(__file__).parent
@@ -74,30 +69,30 @@ if not isExist:
 # load the datafile
 rawdata = loadmat(filepath)
 # obtain the data for training
-data_all, idxgall, posi_all, idxcrop_all = normalize_data(rawdata)
+data_all, idxgall, posi_all, idxcrop_all = normdata_xpinn(rawdata)
 scale = tree_map(lambda x: data_all[x][4][0:2], idxgall)
 
 
 #%% initialization
 
 # initialize the weights and biases of the network
-trained_params = init_xpinns(keys[0], n_hl, n_unit, n_sub=len(idxgall), aniso=True)
+trained_params = init_xpinn(keys[0], n_hl, n_unit, n_sub=len(idxgall), aniso=True)
 
 # create the solution function [tuple(callable, callable)]
-solNN = solu_create(scale)
+solNN = solu_xpinn(scale)
 
 # create the data function for Adam
-dataf = data_sample_create(data_all, idxgall, n_pt)
+dataf = dsample_xpinn(data_all, idxgall, n_pt)
 keys_adam = random.split(keys[1], 5)
 data = dataf(keys_adam[0])
 # create the data function for L-BFGS
-dataf_l = data_sample_create(data_all, idxgall, n_pt2)
+dataf_l = dsample_xpinn(data_all, idxgall, n_pt2)
 key_lbfgs = keys[2]
 
 # group the gov. eqn and bd cond.
 eqn_all = (gov_eqn, front_eqn)
 # calculate the loss function
-NN_loss = loss_aniso_create(solNN, eqn_all, scale, idxgall, lw)
+NN_loss = loss_aniso_xpinn(solNN, eqn_all, scale, idxgall, lw)
 # calculate the initial loss and set it as the loss reference value
 NN_loss.lref = NN_loss(trained_params, data)[0]
 
@@ -118,13 +113,13 @@ wsp_schdul = lambda x: lax.max(10 ** (-x/epoch1 * 2), 0.0125)
 start_time = time.time()
 
 """training with Adam"""
-trained_params, loss1 = adam_optimizer(
+trained_params, loss1 = adam_opt(
     keys_adam[0], NN_loss, trained_params, dataf, epoch1, lr=lr, aniso=True, schdul=wsp_schdul)
 
 # sample the data for L-BFGS training
 data_l = dataf_l(key_lbfgs)
 """training with L-BFGS"""
-trained_params, loss2 = lbfgs_optimizer(NN_loss, trained_params, data_l, epoch2)
+trained_params, loss2 = lbfgs_opt(NN_loss, trained_params, data_l, epoch2)
 
[…]
