+++ paper.md
@@ -63,7 +63,7 @@ viscosity and enhance the training to capture the large-varying viscosity profil
 Due to errors and noise, PINN training with observational data often "cheats", where the networks overfit the training data [@wang2022discovering; @charlie2024euler]. To prevent the issue, the `DIFFICE_jax` package uses residual-based re-sampling scheme during training [@lu2021deepxde; @daw2022mitigating], where more collocation points are sampled in the area with large training residue.
 
 **(5) Extended-PINNs (XPINNs) for large ice shelve:**
-Regular PINN training with a single network cannot capture the rich spatial variation of large ice shelves, such as Ross. The `DIFFICE_jax` package adopts the approach of extended PINNs (XPINNs) [@jagtap2020extended] for studying large ice shelves. This method divides the training domains into several sub-regions, with different networks assigned to each. Detailed description of XPINNs are provided in the `model` folder of the GitHub repository.
+Regular PINN training with a single network cannot capture the rich spatial variation of large ice shelves, such as Ross. The `DIFFICE_jax` package adopts the approach of extended PINNs (XPINNs) [@jagtap2020extended] for studying large ice shelves. This method divides the training domains into several sub-regions, with different networks assigned to each. Detailed description of XPINNs are provided in the `doc` folder of the GitHub repository.
 
 **(6) Inversion of anisotropic viscosity:**
 Prior studies have shown that the viscosity of Antarctica Ice Shelves could be anisotropic [@wang2024deep]. The `DIFFICE_jax` package involves the first algorithm designed to infer anisotropic viscosity. The governing equations, associated boundary conditions, and the cost function for inferring anisotropic viscosity are described in the `docs` folder. 
