### 2024-08-09 #6999 @wangyji
@kthyng also for your previou

> Hi @wangyji and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment over time as I have a chance to go through them:
>
> * [x]  [OSI-approved](https://opensource.org/licenses) license
> * [ ]   Installation instructions
> * [ ]   Documentation
> * [ ]  Tests of some sort
> * [ ]  Check paper (make sure compiles, read through, check for length [250-1000 words](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain))
> * [ ]   [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
> * [x]  Clear research application/in scope
>
> In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers. For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.
>
> Also @wangyji please comment further on the paper you say is simultaneously under review about this software. Can you post a link to it or is it not available publicly? If that is the case, please describe what is in the paper to compare it with this submission.

@kthyng also for your question in the above message, the paper that is currently under review in another journal does not have pre-print version during the review process. That paper is a geoscience paper about Antarctica ice shelves, which aims at elucidating the results inferred by this software in order to explore and uncover the hidden mechanisms that govern the motion of Antarctica ice shelves. However, that paper does not provide detailed algorithms and usage of the software.
