### 2024-08-12 3fd38f5
Update paper.bib
