+++ paper.bib
@@ -21,11 +21,14 @@
 }
 
 @article{cheng2024unified,
-  title={A unified framework for forward and inverse modeling of ice sheet flow using physics-informed neural networks},
+  title={Forward and inverse modeling of ice sheet flow using physics-informed neural networks: Application to Helheim Glacier, Greenland},
   author={Cheng, Gong and Morlighem, Mathieu and Francis, Sade},
-  journal={Authorea Preprints},
+  journal={Journal of Geophysical Research: Machine Learning and Computation},
+  volume={1},
+  number={3},
+  pages={e2024JH000169},
   year={2024},
-  DOI = {10.22541/essoar.170965007.78479393/v1}
+  DOI = {10.1029/2024JH000169}
 }
 
 @article{brinkerhoff2021constraining,
