### 2024-08-15 #6999 @kthyng
@AnjaliSandip you'll need to run this command to assign yourself as editor.
