### 2024-08-15 #6999 @AnjaliSandip
👋 @alextbradley, @pennyhow, @JordiBolibar, @daniel-cheng, & @weiji14 would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:[ https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).
