### 2024-09-13 #6999 @AnjaliSandip
👋 @nmcardoso, @RahulSundar, & @jcla490, would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:[ https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).
