### 2024-09-19 #6999 @RahulSundar
HI!

> 👋 @nmcardoso, @RahulSundar, & @jcla490, would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:[ https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).

I am available to review this submission. Thanks. Would need about 2 weeks for the same.
