### 2024-09-20 #7254 @editorialbot
[REVIEW]: DIFFICE-jax: Differentiable neural-network solver for data assimilation of ice shelves in JAX
**Submitting author:** @wangyji (<a href="http://orcid.org/0000-0002-3987-9038">Yongji Wang</a>)
**Repository:** https://github.com/YaoGroup/DIFFICE_jax
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
