### 2024-11-14 #7254 @RahulSundar
Dear Author @wangyji,

Here are some of my comments.
General feedback: The work and its application to Ice sheet modelling is novel and the code base helps tackle inverse problems of relevant. The paper is well written and crisp. The code base is easily installable without much fuss. The tutorials, examples are well explained. The paper would be a good contribution!

**Specific comments:**
1. Kindly include - automated unit tests for the code base in a separate directory called tests.
2. Include a statement of need in the README.md file
3. Would have been nice to have a sphinx documentation for the code rather than the markdown files.
4. Please include detailed contribution guidelines if possible instead of just the contact email address.

Hope most of these would be possible to implement.
