### 2024-11-17 #7254 @daniel-cheng
Hi @wangjyi, I've been able to complete the review checklist and evaluated this submission.

## General Comments
- This submission provides a new way to perform data assimilation and inverse modeling of ice shelves using Physically Informed Neural Networks (PINNs). The codebase is fully functional, and the provided tutorials/examples serve to document as well as reproduce the results of the submission. The manuscript is well written, and the submission as a whole is well put together, fulfilling most major checklist requirements. After addressing a few technical details noted below, I can recommend acceptance at the editor’s discretion.

## Specific Comments
- There is an unresolved reference on Page 2 Line 45 (?; ?).
- While the tutorial/examples provide a series of manual tests to run in a way, some automated tests could be included. These could be extensions of existing examples, comparing generated results against archived results and ensuring it falls below some threshold.
- Similarly, perhaps the inclusion of pretrained model weights along with results for comparison would aid with validation and automated testing as well as user experience
- Documentation - while present in the form of Readmes covering the files purposes/concepts, and docstrings in the functions themselves - could be made more accessible. Consider an automated documentation generator to scrape and condense the existing docstrings into a more accessible format.
