+++ Contribution.md
@@ -1,0 +1,14 @@
+# 📌 Contributing to DIFFICE-jax
+
+Thank you for considering contributing to `DIFFICE-jax`! We welcome all contributions, whether it’s bug fixes, new features, documentation improvements, or discussions.
+
+---
+
+## 🛠 Getting Started
+
+To start contributing, follow these steps:
+
+1. **Fork the Repository** – Click on the "Fork" button at the top-right corner of this repository.
+2. **Clone Your Fork** – Clone your forked repository to your local machine:
+   ```sh
+   git clone https://github.com/your-username/your-repo.git
