+++ Contribution.md
@@ -3,6 +3,13 @@
 Thank you for considering contributing to `DIFFICE-jax`! We welcome all contributions, whether it’s bug fixes, new features, documentation improvements, or discussions.
 
 ---
+## How do I contribute
+
+### Reporting issues
+
+- **Check Existing Issues** — Pleae first check whether your issue has already been reported. If yes, directly add comments there.
+  
+- **Submit a Report** — If the issue hasn't been reported, open a new issue in the project.
 
 ## 🛠 Getting Started
 
