+++ Contribution.md
@@ -9,7 +9,10 @@ Thank you for considering contributing to `DIFFICE-jax`! We welcome all contribu
 
 - **Check Existing Issues** — Pleae first check whether your issue has already been reported. If yes, directly add comments there.
   
-- **Submit a Report** — If the issue hasn't been reported, open a new issue in the project.
+- **Submit a Report** — If you find a new bug, please open a new issue with.
+  - A clear description of the problem.
+  - Steps to reproduce the issue.
+  - Expected behavior.
 
 ### Suggesting Improvement
 
@@ -19,11 +22,20 @@ If you Have a great idea to improve our package, please open an issue to discuss
 
 We welcome contributions of new datasets, whether for the existing ice shelves in the examples folder or for new ice shelves. Please ensure that your dataset follows the specified [data format](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/data.md). To submit your data, create a Pull Request (see details below).
 
-## 🛠 Getting Started
+---
+## Pull Request
 
-To start contributing, follow these steps:
+To contribute via Pull Request, follow these steps:
+### Getting started
 
 1. **Fork the Repository** – Click on the "Fork" button at the top-right corner of this repository.
 2. **Clone Your Fork** – Clone your forked repository to your local machine:
    ```sh
-   git clone https://github.com/your-username/your-repo.git
+   git clone https://github.com/YaoGroup/DIFFICE_jax.git
+3. **Navigate to the Project Directory**
+   ```sh
+   cd DIFFICE_jax
+4. **Set Up the Environment** – Install dependencies and ensure everything runs smoothly:
+   ```sh
+   pip install -r requirements.txt
+   
