### 2025-02-21 6332943
Create test_func.py
