### 2025-02-21 d85e4b3
Create test_examples.py
