### 2025-02-25 9bacbe3
Add automated unit test
