+++ tests/test_func.py
@@ -1,21 +1,119 @@
-import pymecht as pmt
+import diffice_jax as djax
 import pytest
-import numpy as np
+from jax import random
+import jax.numpy as jnp
+from scipy.io import loadmat
+import math
+import os
 
-def test_arbitrary_MatModel_I1():
-    model = pmt.ARB('mu/2.*(I1-3.)','mu=1.','mu=0.01','mu=10.')
-    mat = pmt.MatModel(model)
-    mat2 = pmt.MatModel('NH')
-    assert (mat.stress(random_F)-mat2.stress(random_F)) == pytest.approx(np.zeros((3,3)))
-    
-def test_arbitrary_MatModel_I2():
-    model = pmt.ARB('c1*(I1-3.) + c2*(I2-3.)','c1=1., c2=1.','c1=0.0001, c2=0.','c1=100., c2=100.')
-    mat = pmt.MatModel(model)
-    mat2 = pmt.MatModel('MR')
-    assert (mat.stress(random_F)-mat2.stress(random_F)) == pytest.approx(np.zeros((3,3)))
-    
-def test_arbitrary_MatModel_J():
-    model = pmt.ARB('kappa/2.*(J-1.)**2','kappa=1.','kappa=1.','kappa=1.')
-    mat = pmt.MatModel(model)
-    mat2 = pmt.MatModel('volPenalty')
-    assert (mat.stress(random_F)-mat2.stress(random_F)) == pytest.approx(np.zeros((3,3)))
+seed = 1234
+key = random.PRNGKey(seed)
+keys = random.split(key, 4)
+
+#%% check the network weight initialization
+@pytest.mark.parametrize("nh, nu, nsub, nblock", [
+    (4, 25, 4, 5),
+    (3, 23, 3, 4),
+    (6, 26, 5, 7),
+])
+
+def test_initialization(nh, nu, nsub, nblock):
+    # test the initialization of regular PINNs
+    trained_params = djax.init_pinn(keys[0], nh, nu)
+    assert isinstance(trained_params, list)
+    assert (len(trained_params)) == pytest.approx(2)
+    assert (len(trained_params[0])) == pytest.approx(nblock)
+    assert (len(trained_params[1])) == pytest.approx(nblock)
+
+    # test the initialization of extended PINNs
+    trained_params2 = djax.init_xpinn(keys[0], nh, nu, nsub)
+    assert len(trained_params2) == pytest.approx(2)
+    assert isinstance(trained_params2, dict)
+    assert "net_u" in trained_params2
+    assert "net_mu" in trained_params2
+    assert len(trained_params2["net_u"]) == nsub
+    assert len(trained_params2["net_mu"]) == nsub
+
+
+#%%
+"""check the normalization of observational data"""
+
+# create the dataset filename
+filename = 'data_pinns_test.mat'
+filepath = os.path.join(os.path.dirname(__file__), filename)
+filename2 = 'data_xpinns_test.mat'
+filepath2 = os.path.join(os.path.dirname(__file__), filename2)
+
+# load the datafile
+rawdata_pinn = loadmat(filepath)
+rawdata_xpinn = loadmat(filepath2)
+data_pinn = djax.normdata_pinn(rawdata_pinn)
+data_xpinn = djax.normdata_xpinn(rawdata_xpinn)
+n_reg = len(rawdata_xpinn['xd'][0])
+
+
+# Fixture with assert
+@pytest.fixture
+def check_normalization():
+    def _check(data_each):
+        # test the data normalization for regular PINNs
+        assert len(data_each[0]) == 2
+        assert len(data_each[1]) == 2
+        assert data_each[2].shape == data_each[3].shape
+        assert len(data_each[4]) == 6
[…]
