+++ docs/source/index.rst
@@ -1,0 +1,17 @@
+.. DIFFICE_jax documentation master file, created by
+   sphinx-quickstart on Tue Feb 25 15:17:44 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+DIFFICE_jax documentation
+=========================
+
+Add your content using ``reStructuredText`` syntax. See the
+`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
+documentation for details.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
