+++ docs/requirements.txt
@@ -1,5 +1,14 @@
-jax == 0.4.26
-jaxlib == 0.4.26
+Sphinx==8.1.3
+sphinx-rtd-theme==3.0.2
+sphinxcontrib-applehelp==2.0.0
+sphinxcontrib-devhelp==2.0.0
+sphinxcontrib-htmlhelp==2.1.0
+sphinxcontrib-jquery==4.1
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==2.0.0
+sphinxcontrib-serializinghtml==2.0.0
+jax==0.4.26
+jaxlib==0.4.26
 chex==0.1.85
 scipy==1.11.4
 numpy==1.26.3
