+++ docs/source/conf.py
@@ -8,16 +8,15 @@
 
 project = 'DIFFICE_jax'
 copyright = '2025, Yongji Wang'
-author = 'Yongji Wang'
+author = 'Yongji Wang & Ching-Yao Lai'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = []
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "sphinx.ext.mathjax", "sphinx.ext.viewcode", "nbsphinx"]
 
 templates_path = ['_templates']
 exclude_patterns = []
-
 
 
 # -- Options for HTML output -------------------------------------------------
