+++ docs/source/conf.py
@@ -5,6 +5,10 @@
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath('../..'))
 
 project = 'DIFFICE_jax'
 copyright = '2025, Yongji Wang'
