+++ docs/source/conf.py
@@ -17,7 +17,7 @@ author = 'Yongji Wang & Ching-Yao Lai'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "sphinx.ext.mathjax", "sphinx.ext.viewcode", "nbsphinx"]
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "sphinx.ext.mathjax", "sphinx.ext.viewcode"]
 
 templates_path = ['_templates']
 exclude_patterns = []
