+++ docs/requirements.txt
@@ -1,13 +1,15 @@
-Sphinx==8.1.3
-sphinx-rtd-theme==3.0.2
-sphinxcontrib-applehelp==2.0.0
-sphinxcontrib-devhelp==2.0.0
-sphinxcontrib-htmlhelp==2.1.0
-sphinxcontrib-jquery==4.1
+# Defining the exact version will make sure things don't break
+sphinx==6.1.3
+sphinx_rtd_theme==1.2.0
+readthedocs-sphinx-search==0.3.2
+sphinx-autodoc-typehints==1.23.0
+sphinxcontrib.applehelp==1.0.4
+sphinxcontrib-devhelp==1.0.2
+sphinxcontrib-htmlhelp==2.0.1
 sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==2.0.0
-sphinxcontrib-serializinghtml==2.0.0
-nbsphinx==5.9.2
+sphinxcontrib-qthelp==1.0.3
+sphinxcontrib-serializinghtml==1.1.5
+nbsphinx==0.9.3
 jax==0.4.26
 jaxlib==0.4.26
 chex==0.1.85
