+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 # Defining the exact version will make sure things don't break
 sphinx==6.1.3
+myst-parser>=3.0.0
 sphinx_rtd_theme==1.2.0
 readthedocs-sphinx-search==0.3.2
 sphinx-autodoc-typehints==1.23.0
