+++ docs/source/conf.py
@@ -8,19 +8,41 @@
 import os
 import sys
 
+# Add the root directory (two levels up) to sys.path
 sys.path.insert(0, os.path.abspath('../..'))
 
+# Add the parent (docs) directory (one level up) to sys.path
+sys.path.insert(0, os.path.abspath('..'))
+
+# Add the tutorial directory (e.g., `tutorial/`)
+sys.path.insert(0, os.path.abspath('../../tutorial'))
+
+# Add the examples directory (e.g., `examples/`)
+sys.path.insert(0, os.path.abspath('../../examples'))
+
 project = 'DIFFICE_jax'
-copyright = '2025, Yongji Wang'
+copyright = '2025, Yongji Wang & Ching-Yao Lai'
 author = 'Yongji Wang & Ching-Yao Lai'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ["sphinx.ext.autodoc", "sphinx.ext.napoleon", "sphinx.ext.mathjax", "sphinx.ext.viewcode", "nbsphinx"]
+extensions = ["myst-parser", 
+              "sphinx.ext.autodoc", 
+              "sphinx.ext.napoleon", 
+              "sphinx.ext.mathjax", 
+              "sphinx.ext.viewcode", 
+              "nbsphinx"]
 
 templates_path = ['_templates']
 exclude_patterns = []
+
+
+# The suffix(es) of source filenames.
+# You can specify multiple suffix as a list of string:
+#
+# source_suffix = ['.rst', '.md']
+source_suffix = ['.rst', '.md']
 
 
 # -- Options for HTML output -------------------------------------------------
