+++ docs/requirements.txt
@@ -1,6 +1,5 @@
 # Defining the exact version will make sure things don't break
 sphinx==6.1.3
-myst-parser>=3.0.0
 sphinx_rtd_theme==1.2.0
 readthedocs-sphinx-search==0.3.2
 sphinx-autodoc-typehints==1.23.0
@@ -11,6 +10,7 @@ sphinxcontrib-jsmath==1.0.1
 sphinxcontrib-qthelp==1.0.3
 sphinxcontrib-serializinghtml==1.1.5
 nbsphinx==0.9.3
+myst-parser==3.0.1
 jax==0.4.26
 jaxlib==0.4.26
 chex==0.1.85
