+++ docs/requirements.txt
@@ -10,7 +10,7 @@ sphinxcontrib-jsmath==1.0.1
 sphinxcontrib-qthelp==1.0.3
 sphinxcontrib-serializinghtml==1.1.5
 nbsphinx==0.9.3
-myst-parser==4.0.1
+myst-parser==2.0.0
 jax==0.4.26
 jaxlib==0.4.26
 chex==0.1.85
