+++ docs/source/index.rst
@@ -9,9 +9,6 @@
 DIFFICE_jax documentation
 =========================
 
-.. include:: ../../README.md
-   :parser: myst_parser
-
 .. automodule:: diffice_jax
    :members:
    :undoc-members:
