### 2025-02-26 d1b0b09
Update and rename docs/Anisotropic.md to docs/source/Anisotropic.md
