### 2025-02-26 40ea05b
Update and rename docs/Isotropic.md to docs/source/Isotropic.md
