### 2025-02-26 9f82353
Update Anisotropic.md
