### 2025-02-26 e6cbc52
Update Anisotropic.md
