+++ docs/source/Anisotropic.md
@@ -54,7 +54,7 @@ $$ \begin{equation}
 \end{equation} $$
 
 where $\mathcal{L_d}$ and $\mathcal{L_e}$ are the data loss and equation loss, respectively, which have the same expression
-as for inferring [isotropic viscosity](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/Isotropic.md). $\gamma_e$ 
+as for inferring [isotropic viscosity](https://github.com/YaoGroup/DIFFICE_jax/blob/main/docs/source/Isotropic.md). $\gamma_e$ 
 and $\gamma_b$ are the weighting pre-factors for the equation and boundary condition loss. Here, we note that the weight 
 $\gamma_g$ for the regularization loss should be set to be much smaller than the weight $\gamma_e$ and $\gamma_b$. Otherwise, 
 the contribution of the regularization loss overwhelms that of the equation loss in the loss function, causing PINNs to first
