### 2025-02-26 2f1ff93
Update Isotropic.md
