### 2025-02-26 3973b58
Update Anisotropic.md
