+++ docs/source/Anisotropic.md
@@ -4,6 +4,8 @@ The conventional Shallow-shelf Approximation (SSA) equations is derived under th
 isotropic properties. However, evidence shows that ice shelves could exhibit anisotropic properties, leading to 
 the viscosity inversion from the remote-sensing data and isotropic SSA equation with no viable solution.
 
+## Governing equations
+
 However, a fully-anisotropic viscosity matrix for ice shelves contains 16 components, making the viscosity
 inversion problem ill-posed with no unique solution. Physically, considering the fact that the non-homogeneous
 structure and crevasse of ice shelves are mainly present in the vertical direction, the viscosity associated 
@@ -20,6 +22,8 @@ $$\begin{eqnarray}
   +\frac{\partial} {\partial x} \left({\mu_h} h \left[\frac{\partial  u}{\partial y}  + \frac{\partial  v}{\partial x} \right]  \right)  = \rho g \left(1-{\frac{\rho}{\rho_w}}\right)h\frac{\partial h}{\partial y}\ .   \qquad 
 \end{eqnarray}$$
 
+## Dynamic boundary conditions
+
 Similarly, the dynamic boundary conditions at the calving front with respect to the anisotropic viscosity components ($\mu_h$, $\mu_v$) becomes
 
 $$\begin{array}{l}
@@ -35,8 +39,8 @@ where $\partial \Omega_c$ indicates the position of ice-shelf calving front, and
 vector towards outwards to the calving front.
 
 It is important to note that the above governing equation and boundary conditions remain insufficient to guarantee 
-the unique inversion of the two viscosity components ($\mu_h$, $\mu_v$). A regularization condition is, thus, required 
-to ensure the uniqueness for inferring the anisotropic viscosity components ($\mu_h$, $\mu_v$). That regularization
+the unique inversion of the two viscosity components ($\mu_h$, $\mu_v$). A **regularization condition** is, thus, required 
+to ensure the uniqueness for inferring the anisotropic viscosity components ($\mu_h$, $\mu_v$). The regularization
 condition we add is that we prefer the solutions of $\mu_h$ and $\mu_v$ are close to each other unless their 
 closeness violates the SSA equations given the data of ice-shelf velocity and thickness. Practically, we can add an 
 regularization term, which measures the mean squared error between the network prediction of $\mu_h$ and $\mu_v$ in 
