### 2025-02-26 827b486
Update Anisotropic.md
