### 2025-02-26 f515732
Update Anisotropic.md
