### 2025-02-26 bf78d3f
Update Anisotropic.md
