### 2025-02-26 115585e
Update Anisotropic.md
