### 2025-02-26 4309e61
Update Isotropic.md
