### 2025-02-26 b8b834d
Update Anisotropic.md
