### 2025-02-26 408f345
Update Anisotropic.md
