### 2025-02-26 e2c9ea7
Create func_eqn.md
