+++ docs/source/func_data.md
@@ -1,3 +1,5 @@
+# Data processing
+
 ## Location: `diffice_jax/data` 
 
 This folder provides the codes that conduct the pre-preprocessing of the input data,
