### 2025-02-26 fd72a81
Create func_models.md
