### 2025-02-26 2e9278e
Update func_models.md
