### 2025-02-26 d3a9433
Update func_models.md
