### 2025-02-26 7e180ec
Update func_models.md
