### 2025-02-26 254cb25
Update func_eqn.md
