### 2025-02-26 2ac82b1
Update func_eqn.md
