### 2025-02-26 d4855ed
Update func_eqn.md
