### 2025-02-26 e01d8db
Create code_optimizer.md
