+++ docs/source/examples.rst
@@ -1,0 +1,7 @@
+.. Examples documentation file, created by
+   sphinx-quickstart on Web Feb 26 15:17:44 2025.
+   You can adapt this file completely to your liking
+
+
+.. include:: ../../examples/README.md
+   :parser: myst_parser.sphinx_
