+++ README.md
@@ -73,7 +73,7 @@ This package is written by Yongji Wang and maintained by Yongji Wang (yongjiw@st
 `DIFFICE_jax` is an open-source software. All code within the project is licensed under the MIT License. For more details, please refer to the [LICENSE](https://github.com/YaoGroup/DIFFICE_jax/tree/main/LICENSE) file.
 
 ## Citation
-Y. Wang, C.-Y. Lai, D. J. Prior and C. Cowen-Breen, Deep learning the flow law of Antarctica ice shelves, *Science*, 2025
+Y. Wang, C.-Y. Lai, D. J. Prior and C. Cowen-Breen, "Deep learning the flow law of Antarctica ice shelves," *Science*, 2025
 ```
 @article{wang2025deep
   title={Deep learning the flow law of Antarctic ice shelves},
