+++ paper.bib
@@ -111,11 +111,13 @@
 }
 
 
-@article{wang2024deep,
+@article{wang2025deep,
   title={Deep learning the flow law of Antarctic Ice Shelves},
   author={Wang, Yongji and Lai, Ching-Yao and and David J. Prior and Cowen-Breen, Charlie},
-  journal = {Under review},
-  year={2024}
+  journal = {Science},
+  year={2025},
+  note={accepted},
+  DOI = {10.1126/science.adp3300}
 }
 
 @article{MacAyeal1989,
@@ -322,4 +324,23 @@
   url = {http://github.com/google/jax},
   version = {0.3.13},
   year = {2018},
+}
+
+@article{rahaman2019spectral,
+  title={On the Spectral Bias of Neural Networks}, 
+  author={Nasim Rahaman and Aristide Baratin and Devansh Arpit and Felix Draxler and Min Lin and Fred A. Hamprecht and Yoshua Bengio and Aaron Courville},
+  journal={arXiv preprint arXiv:1806.08734},
+  year={2019},
+  DOI = {10.48550/arXiv.1806.08734}
+}
+
+@article{xu2020frequency,
+  title={Frequency principle: Fourier analysis sheds light on deep neural networks},
+  author={Xu, Zhi-Qin John and Zhang, Yaoyu and Luo, Tao and Xiao, Yanyang and Ma, Zheng},
+  journal={Communications in Computational Physics},
+  volume={28},
+  number={5},
+  pages={1746-1767},
+  year={2020},
+  DOI={10.4208/cicp.OA-2020-0085}
 }
