### 2025-02-27 9269fbc
Update code_eqn.md
