### 2025-02-27 68f0a9d
Update MANIFEST.in
