+++ MANIFEST.in
@@ -2,3 +2,4 @@ exclude *.md
 exclude examples/*
 exclude docs/*
 exclude tutorial/*
+exclude tests/*
