### 2025-02-27 #7254 @wangyji
@AnjaliSandip Thank you for waiting. It took a bit time for me to learn how to use the sphinx documentation.

I am writing to confirm that I have addressed the comments from the reviewers, which include:

1. Adding a test folder which include automated unit tests.

2. Add the statement of need in the README.file of the repository

3. Using sphinx documentation (create a Readthedocs website) to better organize the documents.

4. Add a pertained model in the tutorial examples for users to validate the results

5. fix the reference issue in the previous paper (Page 2 Line 45)

Please let me know how to resubmit the paper for the review process or perhaps you could help do that for me.
