### 2025-02-27 #7254 @wangyji
missing one point:

point 6: I add a detailed contribution guideline in the repository
