### 2025-04-08 9c926e3
update the tutorial examples
