+++ tutorial/COMSOL/txt2mat.m
@@ -1,5 +1,5 @@
 clear;
-close all;
+% close all;
 
 % type the filename of data to convert
 filename = 'SynData_exp1';
@@ -10,34 +10,39 @@ DataPath = [filename, '.txt'];
 % read the data file from COMSOL
 a = readtable(DataPath);
 
+% get the first variable in the table
+x = a.Var1;
+% find the first non-Nan value of x
+idx = find(~isnan(x),1,'first');
 % remove the description lines at the top of the data file
-a(1:7, :) = [];
+a(1:idx-1,:) = [];
 
-% extract each variable from the data file
+% extract each variable from the data file (1D array)
 x = a.Var1;
 y = a.Var2;
 u = a.Var3;
 v = a.Var4;
 h = a.Var5;
 mu = a.Var6;
+eb1 = a.Var7;
+eb2 = a.Var8;
 
-% define the equally-spaced grids in each dimension
-x0 = linspace(0,max(x),401);
-y0 = linspace(0,max(y),301);
-[xd, yd] = meshgrid(x0, y0);
+% set the shape of the 2D matrix
+% (consistent with the grid numbers set in the COMSOL->Export->data)
+Shape2D = [400, 250];
 
-% define another equally-spaced grids in each dimension
-x1 = linspace(0,max(x),301);
-y1 = linspace(0,max(y),201);
-[xd_h, yd_h] = meshgrid(x1, y1);
+% convert the 1D-array data to 2D matrix
+xd = reshape(x, Shape2D).';
+yd = reshape(y, Shape2D).';
+ud = reshape(u, Shape2D).';
+vd = reshape(v, Shape2D).';
+hd = reshape(h, Shape2D).';
+mud = reshape(mu, Shape2D).';
+eb1d = reshape(eb1, Shape2D).';
+eb2d = reshape(eb2, Shape2D).';
 
-% interpolate the data in terms of the given grids
-ud = griddata(x,y,u,xd,yd);
-vd = griddata(x,y,v,xd,yd);
-mud = griddata(x,y,mu,xd,yd);
-
-% interpolate the thickness data on another grids
-hd = griddata(x,y,h,xd_h,yd_h);
+xd_h = xd;
+yd_h = yd;
 
 % plot the interpolated data before saving
 figure; surf(xd,yd,ud);
@@ -45,8 +50,8 @@ shading interp;
 colormap jet
 
 % set the positions for the calving front
-xct0 = xd(:, end-3:end);
-yct0 = yd(:, end-3:end);
+xct0 = xd(:, end-1:end);
+yct0 = yd(:, end-1:end);
 xct = reshape(xct0.', [numel(xct0), 1]);
 yct = reshape(yct0.', [numel(yct0), 1]);
 % set the associated normal vector to the calving front
