### 2025-04-08 #7254 @wangyji
Hi @AnjaliSandip, here is the Zenodo address for the latest release.
https://doi.org/10.5281/zenodo.15175676

Please let me know if you need any other information
