### 2025-04-11 #7254 @wangyji
Hi @AnjaliSandip just want to check the status of the paper. Is there anything I need to do before the official acceptance of the paper. Thanks a lot
