### 2025-04-11 #7254 @AnjaliSandip
> Hi [@AnjaliSandip](https://github.com/AnjaliSandip) just want to check the status of the paper. Is there anything I need to do before the official acceptance of the paper. Thanks a lot

None that I can think of.
