### 2025-04-11 #7254 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
