### 2025-04-14 #7254 @wangyji
@AnjaliSandip Thank you for your reply. I just wanted to check whether the paper has been officially accepted. Also, do you know how long it usually takes for it to be published on the journal's website?
