### 2025-04-21 #7254 @wangyji
@kthyng Just confirm that I have update the version of the software and make a new release. The current version number is updated to be v1.0.0
