### 2025-04-25 0b1873f
Update paper.bib
