+++ paper.bib
@@ -21,7 +21,7 @@
 }
 
 @article{cheng2024unified,
-  title={Forward and inverse modeling of ice sheet flow using physics-informed neural networks: Application to Helheim Glacier, Greenland},
+  title={Forward and inverse modeling of ice sheet flow using physics-informed neural networks: Application to {H}elheim {G}lacier, {G}reenland},
   author={Cheng, Gong and Morlighem, Mathieu and Francis, Sade},
   journal={Journal of Geophysical Research: Machine Learning and Computation},
   volume={1},
@@ -32,7 +32,7 @@
 }
 
 @article{brinkerhoff2021constraining,
-  title={Constraining subglacial processes from surface velocity observations using surrogate-based Bayesian inference},
+  title={Constraining subglacial processes from surface velocity observations using surrogate-based {B}ayesian inference},
   author={Brinkerhoff, Douglas and Aschwanden, Andy and Fahnestock, Mark},
   journal={Journal of Glaciology},
   volume={67},
@@ -74,7 +74,7 @@
 }
 
 @article{jouvet2023inversion,
-  title={Inversion of a Stokes glacier flow model emulated by deep learning},
+  title={Inversion of a {S}tokes glacier flow model emulated by deep learning},
   author={Jouvet, Guillaume},
   journal={Journal of Glaciology},
   volume={69},
@@ -103,7 +103,7 @@
 }
 
 @article{wang2022discovering,
-  title={Discovering the rheology of Antarctic Ice Shelves via physics-informed deep learning},
+  title={Discovering the rheology of {A}ntarctic ice shelves via physics-informed deep learning},
   author={Wang, Yongji and Lai, Ching-Yao and Cowen-Breen, Charlie},
   journal = {SSRN Prepreint},
   year={2022},
@@ -112,7 +112,7 @@
 
 
 @article{wang2025deep,
-  title={Deep learning the flow law of Antarctic Ice Shelves},
+  title={Deep learning the flow law of {A}ntarctic ice shelves},
   author={Wang, Yongji and Lai, Ching-Yao and and David J. Prior and Cowen-Breen, Charlie},
   journal = {Science},
   year={2025},
@@ -122,7 +122,7 @@
 
 @article{MacAyeal1989,
   author = {MacAyeal, Douglas R.},
-  title = {Large-scale ice flow over a viscous basal sediment: Theory and application to ice stream B, Antarctica},
+  title = {Large-scale ice flow over a viscous basal sediment: {T}heory and application to ice stream {B}, {A}ntarctica},
   journal = {Journal of Geophysical Research: Solid Earth},
   volume = {94},
   pages = {4071-4087},
@@ -154,7 +154,7 @@
 
 
 @article{ranganathan2024modified,
-	title={A modified viscous flow law for natural glacier ice: Scaling from laboratories to ice sheets},
+	title={A modified viscous flow law for natural glacier ice: {S}caling from laboratories to ice sheets},
 	author={Ranganathan, Meghana and Minchew, Brent},
 	journal={Proceedings of the National Academy of Sciences},
 	volume={121},
@@ -165,7 +165,7 @@
 }
 
 @article{furst2015assimilation,
-	title={Assimilation of Antarctic velocity observations provides evidence for uncharted pinning points},
+	title={Assimilation of {A}ntarctic velocity observations provides evidence for uncharted pinning points},
 	author={F{\"u}rst, JJ and Durand, G and Gillet-Chaulet, F and Merino, N and Tavard, L and Mouginot, J and Gourmelen, N and Gagliardini, O},
[…]
