+++ paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{perego2014optimal,
-  title={Optimal initial conditions for coupling ice sheet models to Earth system models},
+  title={Optimal initial conditions for coupling ice sheet models to {E}arth system models},
   author={Perego, Mauro and Price, Stephen and Stadler, Georg},
   journal={Journal of Geophysical Research: Earth Surface},
   volume={119},
@@ -154,7 +154,7 @@
 
 
 @article{ranganathan2024modified,
-	title={A modified viscous flow law for natural glacier ice: {S}caling from laboratories to ice sheets},
+	title={A modified viscous flow law for natural glacier ice: Scaling from laboratories to ice sheets},
 	author={Ranganathan, Meghana and Minchew, Brent},
 	journal={Proceedings of the National Academy of Sciences},
 	volume={121},
