### 2025-04-25 #7254 @wangyji
Hi @kthyng, thank you for the reminder. Just confirm that I have doubled checked the capitalization in our references and updated the .bib file accordingly. Please let me know if you find any other issues.
