### 2025-04-30 #7254 @kthyng
@openjournals/dev Looks like this is the error, any ideas?

```
To use retry middleware with Faraday v2.0+, install `faraday-retry` gem
/home/runner/work/_actions/xuanxu/upload-files-action/main/vendor/bundle/ruby/3.3.0/gems/octokit-8.1.0/lib/octokit/response/raise_error.rb:14:in `on_complete': PUT https://api.github.com/repos/***nals/joss-papers/contents/joss.07254/10.21105.joss.07254.pdf: 422 - Invalid request. (Octokit::UnprocessableEntity)

"sha" wasn't supplied. // See: https://docs.github.com/rest/repos/contents#create-or-update-file-contents
```
